% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMetricToSeObj.R
\name{addMetricToSeObj}
\alias{addMetricToSeObj}
\title{Adds results and plots of a metric to SummarizedExperiment object.}
\usage{
addMetricToSeObj(
  se.obj,
  slot = "Metrics",
  assay.names,
  assessment.type = "gene.level",
  assessment,
  method,
  variables,
  file.name,
  results.data
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{slot}{Character. The name of the slot in the \code{metadata} of the \code{SummarizedExperiment} object.}

\item{assay.names}{Character or character vector. The name(s) of the assay(s) in the \code{SummarizedExperiment} object.
The default is set to 'all'.}

\item{assessment.type}{Character. The type of assessment to perform. Options are 'gene.level' or 'global.level'.}

\item{assessment}{Character. The name of the metric to be evaluated.}

\item{method}{Character. The method used to calculate the metric.}

\item{variables}{Character or character vector. The variable(s) used to calculate the metric.}

\item{file.name}{Character. The name of the file to which the metric results will be written.}

\item{results.data}{Character. The name of the object or file containing the metric results.}
}
\description{
Adds results and plots of a metric to SummarizedExperiment object.
}
\author{
Ramyar Molania
}
