% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOverallPlotToSeObj.R
\name{addOverallPlotToSeObj}
\alias{addOverallPlotToSeObj}
\title{Add overall plot to the "metadata" of SummarizedExperiment object.}
\usage{
addOverallPlotToSeObj(
  se.obj,
  slot = "Plots",
  assessment.type = "gene.level",
  assessment,
  method,
  variables,
  file.name,
  plot.data
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{slot}{Character A character string  indicating the name of the slots in the 'metadata' of the SummarizedExperiment object.}

\item{assessment.type}{Character A character string  indicating the type of assessment. Options are \code{gene.level} or
\code{global.level}.}

\item{assessment}{Character A character string  indicating the name of the metric to be checked.}

\item{method}{Character A character string  indicating the method used to calculate the metric.}

\item{variables}{Character A character string or a vector of symbols indicating the variables used to calculate the metric.}

\item{file.name}{Character A character string indicating the file name to which the results of the metric are assigned.}

\item{plot.data}{Character A character string indicating the name of the plots file for the metric.}
}
\description{
Add overall plot to the "metadata" of SummarizedExperiment object.
}
\author{
Ramyar Molania
}
