% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTcgaBatchInfo.R
\name{addTcgaBatchInfo}
\alias{addTcgaBatchInfo}
\title{Adds the TCGA RNA-seq batch information}
\usage{
addTcgaBatchInfo(
  se.obj,
  use.lables = FALSE,
  batch.info = "all",
  cancer.type = NULL,
  order.se.obj = TRUE,
  factors.to.order = batch.info,
  missing.samples.name = "DF"
)
}
\arguments{
\item{se.obj}{TTT}

\item{use.lables}{TTT}

\item{batch.info}{TTTT}

\item{cancer.type}{TTTTsss}

\item{order.se.obj}{TTTT}

\item{factors.to.order}{TTTT}

\item{missing.samples.name}{TTTTT}
}
\description{
This function adds the TCGA RNA-seq batch information to the a SummarizedExperiment object.
}
\author{
Ramyar Molania
}
