% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTcgaClinicalInfo.R
\name{addTcgaClinicalInfo}
\alias{addTcgaClinicalInfo}
\title{Add the TCGA RNA-seq batch information}
\usage{
addTcgaClinicalInfo(
  se.obj,
  tissue.type,
  factors = c("OS", "OS.time"),
  cancer.type = NULL,
  missing.samples.name = "DF",
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A \code{SummarizedExperiment} object containing TCGA RNA-seq data.}

\item{tissue.type}{A data frame or character string specifying batch information. This can be a metadata table
with batch IDs for each sample, or a file path to such a table.}

\item{factors}{Character. A TCGA cancer type code (e.g., \code{"BRCA"}, \code{"PRAD"}, \code{"COAD"}) used to match batch
annotations with the corresponding dataset.}

\item{cancer.type}{Logical. If \code{TRUE}, all samples in the \code{SummarizedExperiment} object are retained, even if
batch information is missing. If \code{FALSE}, samples without batch annotations are removed. The default is \code{TRUE}.}

\item{missing.samples.name}{Character. A label to assign to samples without batch information. The default is
\code{"Unknown"}.}

\item{verbose}{Logical. Indicates whether to display output messages during function execution. The default is set to
\code{TRUE}.}
}
\value{
A \code{SummarizedExperiment} object with TCGA RNA-seq batch information added to the sample annotation (\code{colData}) of
\code{SummarizedExperiment} object.
}
\description{
This function adds the TCGA RNA-seq batch information to a \code{SummarizedExperiment} object, using metadata
provided by TCGA sample barcodes or external batch information files.
}
\author{
Ramyar Molania
}
