% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyOtherNormalizations.R
\name{applyOtherNormalizations}
\alias{applyOtherNormalizations}
\title{Performs several normalization methods for RNA-seq data.}
\usage{
applyOtherNormalizations(
  se.obj,
  assay.name,
  method = "CPM",
  apply.log = TRUE,
  pseudo.count = 1,
  check.se.obj = TRUE,
  remove.na = "assays",
  new.assay.name = NULL,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{Character. A character that specifies the name of the data (assay) in the SummarizedExperiment object.
The selected data should contain raw count data.}

\item{method}{Character. A character that indicates which normalization method should be applied the data sets specified
by \code{assay.name}.
Options include:
\itemize{
\item \code{'CPM'}: Counts Per Million, from the \strong{edgeR} package.
\item \code{'TMM'}: Trimmed Mean of M-values, from the \strong{edgeR} package.
\item \code{'upper'}: A scaling normalization that adjusts the upper quartile of each lane, from the \strong{EDASeq} package.
\item \code{'median'}: A scaling normalization that adjusts the median of each lane, from the \strong{EDASeq} package.
\item \code{'full'}: Full quantile normalization (nonlinear), from the \strong{EDASeq} package.
\item \code{'VST'}: Variance Stabilizing Transformation, from the \strong{DESeq2} package.
The default is set to \code{'CPM'}.
}}

\item{apply.log}{Logical. Indicates whether to apply a log transformation to the data after normalization. The default
is set to \code{TRUE}.}

\item{pseudo.count}{Numeric. A numeric value as pseudo count to be added to all measurements after normalization and
before log transformation to avoid \code{-Inf} for zero values. The default is set to \code{1}.}

\item{check.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object or not before applying
nationalizations. The default is set  to \code{TRUE}. We refer to the \code{checkSeObj()} function for more details.}

\item{remove.na}{Character. A character that indicates whether to remove NA or missing values from the data sets or
not. The options are \code{'assays'} or \code{'none'}. The default is set to \code{'assays'}. Refer to the \code{checkSeObj()} function
for more details.}

\item{new.assay.name}{Character. A character that specifies the name of the newly normalized assay. If set to \code{NULL}
(default), the function will automatically generate a name using the format: paste0(method, ' on ', assay.name).}

\item{save.se.obj}{Logical. Indicates whether to save the normalized data as new data (assay) within the
SummarizedExperiment object. If set to \code{FALSE}, the normalized data will be returned as a matrix. The default is set
to \code{TRUE}. Then, each normalized data will be added a new data (assay) to SummarizedExperiment object.}

\item{verbose}{Logical. Indicates whether to display output messages during function execution. The default is set to
\code{TRUE}.}
}
\value{
Either a SummarizedExperiment object containing the newly normalized data, or a normalized expression matrix
if \code{save.se.obj} is set to \code{FALSE}.
}
\description{
This function provides several normalization methods for RNA-seq data, including CPM, TMM,  upper-quartile, median,
full quantile (nonlinear), and VST. It utilizes the \strong{EDASeq}, \strong{DESeq2}, and \strong{edgeR} R packages for normalization.
Note that the default parameters of these underlying functions are used.
}
\author{
Ramyar Molania
}
