% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessNCGs.R
\name{assessNCGs}
\alias{assessNCGs}
\title{Assess the performance of NCG.}
\usage{
assessNCGs(
  se.obj,
  assay.name,
  variables.to.assess.ncg,
  ncg,
  apply.log = TRUE,
  pseudo.count = 1,
  nb.pcs = 3,
  svd.bsparam = bsparam(),
  center = TRUE,
  scale = FALSE,
  plot.output = TRUE,
  check.se.obj = TRUE,
  remove.na = "none",
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{Character. A character specifying the name of the data (assay) in the SummarizedExperiment object
to be selected. This data be will be used to assess the performance of NCGs. The data should the same one used for the
identification of the NCG.}

\item{variables.to.assess.ncg}{Character or character vector. Specifies the column name(s) in the sample annotation
of the SummarizedExperiment object corresponding to variables for which NCG performance should be assessed.}

\item{ncg}{Logical. A logical vector specifying a set of NCG whose performance will be assessed.}

\item{apply.log}{Logical. Indicates whether to apply log2 transformation to the selected data set. The default is set
to \code{TRUE}.}

\item{pseudo.count}{Numeric. A numeric value to be added to all measurements in the selected
data sets before applying the log2 transformation to avoid \code{-Inf} values for zero measurements. The default is set to 1.}

\item{nb.pcs}{Numeric. Specifies the number of principal components to retain when performing PCA on the data. The PCs
will be used for the assessment. The default is set to 3.}

\item{svd.bsparam}{Character. A BiocParallelParam object specifying how palatalization should be performed for performing
PCA using SVD. The default is set to \code{bsparam()}. We refer to the \code{runSVD()} function from the \strong{BiocSingular} R package
for further details.}

\item{center}{Logical. Indicates whether to center the data or not before calculating PCs. If center is \code{TRUE}, then
centering is done by subtracting the column means of the assay from their corresponding columns. The default is set
to \code{TRUE}.}

\item{scale}{Logical. Indicates whether to scale the data or not before calculating PCs. If scale is set to \code{TRUE}, then
scaling is done by dividing the (centered) columns of the assays by their standard deviations if center is \code{TRUE}, and
the root mean square otherwise. The default is set to \code{FALSE}.}

\item{plot.output}{Logical. Indicates whether to generate and display performance plots of NCG. Default is \code{TRUE}.}

\item{check.se.obj}{Logical. Indicates whether to validate the SummarizedExperiment object before analysis. The default
is set to \code{TRUE}.}

\item{remove.na}{Character. Indicates whether to remove NA or missing values from the data sets. Options are: \code{assays}
or \code{none}. Default is \code{assays}. Refer to the \code{checkSeObj()} function for more details.}

\item{verbose}{Logical. Indicates whether to display messages during function execution. The default is set \code{TRUE}.}
}
\value{
A line-dot plot show the association between the first \code{nb.pcs} PCs and the variables specified by \code{variables.to.assess.ncg}.
}
\description{
Assess the performance of NCG.
}
\author{
Ramyar Molania
}
