% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessNormalization.R
\name{assessNormalization}
\alias{assessNormalization}
\title{Assess the performance of RNA-seq normalization methods.}
\usage{
assessNormalization(
  se.obj,
  assay.names = "all",
  bio.variables,
  uv.variables,
  assessment.level = "L1",
  assessments.to.exclude = NULL,
  select.top.ruv = TRUE,
  fast.pca = TRUE,
  sil.dist.measure = "euclidian",
  sli.nb.pcs = 3,
  ari.clustering.method = "hclust",
  ari.hclust.method = "complete",
  ari.hclust.dist.measure = "euclidian",
  ari.nb.pcs = 3,
  corr.method = "spearman",
  corr.cutoff = 0.2,
  anova.method = "aov",
  fvalue.cutoff = 1,
  vca.nb.pcs = 3,
  lra.nb.pcs = 3,
  pcorr.method = "spearman",
  pcorr.cutoff = 0.3,
  bio.weight = 0.6,
  uv.weight = 0.4,
  plot.output = TRUE,
  save.se.obj = TRUE,
  output.name = NULL,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A \code{SummarizedExperiment} object.}

\item{assay.names}{Character. A character string or a vector of character strings for selecting the name(s) of the
assay(s) in the \code{SummarizedExperiment} object to compute PCA. The default is set to \code{all}, which indicates all the
assays in the \code{SummarizedExperiment} object will be selected.}

\item{bio.variables}{Character or character vector. One or more column names of the sample annotation indicating known
biological variables (categorical, continuous or both) in the \code{SummarizedExperiment} object.}

\item{uv.variables}{character or character vector. One or more column names indicating unwanted variables
(categorical or continuous) in the SummarizedExperiment object.}

\item{assessment.level}{A character string specifying the assessment level. Options are \code{L1} and \code{L2}. The default
is set to \code{L1}. The two levels differ in the number of normalization assessment metrics applied: \code{L2} uses the full
set of assessment metrics implemented in \pkg{RUVprps}, whereas \code{L1} uses a reduced subset of these metrics.}

\item{assessments.to.exclude}{A character or character vector. Names of assessment metrics to exclude, as returned by
the \code{getAssessmentMetrics()} function. Default is NULL.}

\item{select.top.ruv}{A logical indicating whether to return only the top-performing RUV-III–normalized dataset based
on the final assessment. If \code{TRUE}, only the best-performing normalized dataset is returned in the final assessment
table. If \code{FALSE}, all RUV-III–normalized datasets are returned in the \code{SummarizedExperiment} object.}

\item{fast.pca}{Logical. Whether to use fast PCA. The default is to \code{TRUE}.}

\item{sil.dist.measure}{character. Distance measure to use for silhouette analysis. Options: 'euclidean', 'maximum',
'manhattan', 'canberra', 'binary', or 'minkowski'. Default is 'euclidean'.}

\item{sli.nb.pcs}{Numeric. Number of principal components to use for silhouette analysis. Default is 3. Must not exceed
\code{compute.nb.pcs}.}

\item{ari.clustering.method}{character. Clustering method for ARI computation. Options: 'mclust' or 'hclust'. Default
is 'hclust'.}

\item{ari.hclust.method}{character. Method used in \code{hclust} for ARI. Options: 'ward.D', 'ward.D2', 'single', 'complete',
'average', 'mcquitty', 'median', or 'centroid'. Default is 'complete'.}

\item{ari.hclust.dist.measure}{character. Distance measure for \code{hclust} in ARI analysis. Options:'euclidean', 'maximum',
'manhattan', 'canberra', 'binary', or 'minkowski'. Default is 'euclidean'.}

\item{ari.nb.pcs}{Numeric. Number of principal components to use for ARI analysis. Default is 3. Must not exceed
\code{compute.nb.pcs}.}

\item{corr.method}{character. Correlation method for gene-variable correlation. Options: 'spearman' or 'pearson'.
Default is 'spearman'.}

\item{corr.cutoff}{Numeric. Threshold for selecting genes with an absolute correlation lower than this value with
continuous variables. The default is set to 0.2.}

\item{anova.method}{character. ANOVA method for gene-variable comparison. Options: 'aov' or 'welch'. Default is 'aov'.}

\item{fvalue.cutoff}{Numeric. Threshold for selecting genes with an F-statistic lower than this value with categorical
variables. Te default is set to 1.}

\item{vca.nb.pcs}{Numeric. Number of principal components to use for variance component analysis. Default is 3. Must
not exceed \code{compute.nb.pcs}.}

\item{lra.nb.pcs}{Numeric. Number of principal components to use for latent representation analysis. Default is 3.
Must not exceed \code{compute.nb.pcs}.}

\item{pcorr.method}{character. Correlation method for gene-gene ordinary and partial correlations. Options: 'spearman'
or 'pearson'. The default is set to \code{spearman}.}

\item{pcorr.cutoff}{Numeric. Threshold for selecting gene-gene correlations where the difference between ordinary and
partial correlations is less than this value.}

\item{bio.weight}{Numeric. Weight for the biological preservation score. Default is 0.6. See details for more information.}

\item{uv.weight}{Numeric. Weight for the removal of unwanted variation score. Default is 0.4. See details for more
information.}

\item{plot.output}{Logical. If \code{TRUE}, displays the final assessment plot.}

\item{save.se.obj}{Logical. If \code{TRUE}, saves the results in the SummarizedExperiment object. Default is set to \code{TRUE}.}

\item{output.name}{A character. File name for the results stored in the SummarizedExperiment. If \code{NULL}, the function creates
a default name using \code{paste0('RUVprps_assessNormalization', length(assays(se.obj)), '_assays.')}.}

\item{verbose}{Logical. If \code{TRUE}, displays messages during function execution.}
}
\value{
A \code{SummarizedExperiment} object containing the assessment matrix, plot, and table, or a list containing all results.
}
\description{
This function summarizes a range of gene-level and global metrics obtained using the \code{assessVariation} function for
individual biological and unwanted variables. The function returns numerical assessments in a table to assess the
performance of difference normalization. Refer to details for more information.
}
\details{
It is essential to assess the performance of a normalization as a remover of unwanted variation and a preserver of
biological variation in the data. An ideal normalization should preserve all known sources of biological variation
and reduce or remove the impact of all known sources of unwanted variation. To rank methods according to their
performance, we aggregate all evaluation metrics into a overall score as follows.
Several assessment will be performed:
For each categorical variable:
\itemize{
\item PCA plot of the categorical variable.
\item Silhouette, ARI, LISI, kBET computed on categorical variables.
\item Differential analysis based ANOVA between the gene expression and the categorical variable.
\item Vector correlation between the first cumulative PCs of the gene expression and the categorical variable.
For each continuous variable:
\item Linear regression between the first several (cumulatively) PCs and continuous variable.
\item Partial and ordinary correlation between gene expression and continuous variables separately.
}

It will output the following plots:
\itemize{
\item PCA plot of each categorical variable.
\item Boxplot of the F-test distribution from ANOVA between the gene expression and each categorical variable.
\item Vector correlation between the first cumulative PCs of the gene expression and each categorical variable.
\item Combined Silhouette plot of the combined pair of all categorical variables.
\item Linear regression between the first cumulative PC and continuous variable.
\item Boxplot of the correlation between gene expression and continuous variable.
\item It will also output the RLE plot distribution.
}
}
\references{
**Molania R., ..., Speed, T. P., A new normalization for Nanostring nCounter gene expression data, Nucleic Acids
Research, 2019.
**Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
\author{
Ramyar Molania
}
