% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeLmmFormula.R
\name{changeLmmFormula}
\alias{changeLmmFormula}
\title{Changes the formula for linear mixed model}
\usage{
changeLmmFormula(form, out.put = "character", sub.set = NULL)
}
\arguments{
\item{form}{TTT}

\item{out.put}{TTTT}

\item{sub.set}{TTTT}
}
\description{
Changes the formula for linear mixed model
}
