% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPRPSconnectedness.R
\name{checkPRPSconnectedness}
\alias{checkPRPSconnectedness}
\title{Assess the connectedness across PRPS sets.}
\usage{
checkPRPSconnectedness(data.input, min.samples, batch.name, verbose = TRUE)
}
\arguments{
\item{data.input}{A matrix (table) of PRPS sets across all batches.}

\item{min.samples}{Numeric. The minimum number of samples required to create a pseudo-sample.}

\item{batch.name}{Character. A character string indicating the name of the source of unwanted variation for which the
connectivity between PRPS sets is assessed.}

\item{verbose}{Logical. If \code{TRUE}, the function displays progress messages. The default is \code{TRUE}.}
}
\description{
A replicate or pseudo-replicate set is said to span two batches if it includes samples  or pseudo-samples from both
batches. In such cases, RUV-III will remove the batch  differences captured by those replicates across the two batches.
For experiments with multiple batches, it may not be possible to have a single replicate or pseudo-replicate set that
spans all of them. Instead, it is necessary to construct replicate or pseudo-replicate sets with connections that collectively
cover all batches. The functions that create PRPS data include the argument \code{check.prps.connectedness}.
When set to \code{TRUE}, the function checks connectivity between PRPS sets and reports  how they collectively cover
the specified batches.
}
\author{
Ramyar Molania
}
