% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSeObj.R
\name{checkSeObj}
\alias{checkSeObj}
\title{Assesses SummarizedExperiment object.}
\usage{
checkSeObj(
  se.obj,
  assay.names = "all",
  variables = "all",
  remove.na = "both",
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Character or character vector. A character or character vector of the name(s) of the data(assay) in
the SummarizedExperiment object. The default is set to \code{all}, indicating that all data in the SummarizedExperiment
object will be assessed.}

\item{variables}{Character. A character or character vectors of the name(s) of the column(s) in the sample annotation
of the SummarizedExperiment object. The default it is set to \code{all}, in which case all columns will be examined. We
recommend specifying only those column(s) that are of interest to your analysis.}

\item{remove.na}{Character. A character string that specifies whether to eliminate missing values from \code{assays},
\code{sample.annotation}, \code{both}, or \code{none}. When \code{assays} is chosen, genes containing missing values will be omitted. If
\code{sample.annotation} is selected, samples with NA or missing values for each \code{variables} will be excluded. The default
is set to \code{both}.}

\item{verbose}{Logical. If \code{TRUE}, shows the messages of different steps of the function.}
}
\value{
A SummarizedExperiment object.
}
\description{
This function assesses the structure of a SummarizedExperiment object and removes any missing values from both
data(assay) and sample annotation.
}
\details{
When multiple assays are provided, if data are missing in only one of the datasets, the corresponding rows in the other
datasets will also be removed. If the specified variables contain missing values, the corresponding samples will be removed
from the data as well. Please note that the current RUV-III method does not support missing values in the data.
}
\author{
Ramyar Molania
}
