% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeARI.R
\name{computeARI}
\alias{computeARI}
\title{Compute Adjusted Rand Index (ARI).}
\usage{
computeARI(
  se.obj,
  assay.names = "all",
  variable,
  clustering.method = "hclust",
  hclust.method = "complete",
  hclust.dist.measure = "euclidian",
  fast.pca = TRUE,
  nb.pcs = 3,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{character or character vector. A character or a vector of characters of name(s) of the data(assay)
in the SummarizedExperiment object to use for ARI computation. The default is set to \code{all}, meaning all assays in the
object will be selected.}

\item{variable}{character. The name of the column containing the categorical variable in the SummarizedExperiment object.
This variable can represent either a biological or an unwanted factor.}

\item{clustering.method}{character. A character that indicates which clustering method to be applied on the principal
components to calculate the ARI. Options are \code{mclust} or \code{hclust}. The default is se to the \code{hclust} method.}

\item{hclust.method}{Character. Agglomeration method to use when \code{clustering.method} is set to \code{hclust}. Options include:
\code{ward.D}, \code{ward.D2}, \code{single}, \code{complete}, \code{average} (= UPGMA), \code{mcquitty} (= WPGMA), \code{median} (= WPGMC), or \code{centroid}
(= UPGMC). See the \code{hclust} function in the \strong{stats} package for more details.}

\item{hclust.dist.measure}{Character. A character string specifying which distance measure to be used in the \code{dist} function
when applying hierarchical clustering. Options are: \code{euclidean}, \code{maximum}, \code{manhattan}, \code{canberra}, \code{binary}, or
\code{minkowski}. See the \code{dist} function in the \strong{stats} package for more details.}

\item{fast.pca}{Logical. Indicates whether to use principal components computed via fast PCA by the \code{comptePCA} function
The default is set to \code{TRUE}. Note that using fast PCA or standard PCA does not affect silhouette coefficient calculation.
See details for more information.}

\item{nb.pcs}{Numeric. Number of first principal components to use when calculating distances between samples. The efault
is to 3.}

\item{save.se.obj}{Logical. If \code{TRUE}, saves the results in the metadata of the SummarizedExperiment object; otherwise,
returns the result as list. The default is set to \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE}, displays messages for each step of the function.}
}
\value{
A SummarizedExperiment object or a list containing the computed ARI for specified categorical variable.
}
\description{
This functions computes the adjusted rand index(ARI) for given a categorical variable using the first n PCs of the
specified dataset(s) in a SummarizedExperiment object.
}
\details{
The ARI is the corrected-for-chance version of the Rand index. The ARI measures the percentage of matches between
two label lists. We use the ARI to assess the performance of normalization methods in terms of sample subtype
separation and batch mixing. We first calculate PCs and use the first PCs to perform ARI.
}
\references{
Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
\author{
Ramyar Molania
}
