% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGenesVariableTwoWayAnova.R
\name{computeGenesVariableTwoWayAnova}
\alias{computeGenesVariableTwoWayAnova}
\title{Computes ANOVA between individual gene expression and a categorical variable}
\usage{
computeGenesVariableTwoWayAnova(
  se.obj,
  assay.names,
  bio.variables,
  uv.variables,
  nb.bio.clusters = 3,
  bio.clustering.method = "kmeans",
  nb.uv.clusters = 3,
  uv.clustering.method = "kmeans",
  samples.to.use = "all",
  apply.log = TRUE,
  pseudo.count = 1,
  nb.cores = 1,
  check.se.obj = TRUE,
  remove.na = "none",
  override.check = FALSE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A \code{SummarizedExperiment} object.}

\item{assay.names}{Character. A character string or vector specifying the name(s) of the assay(s) to compute ANOVA.
Defaults to \code{"all"} to select all assays.}

\item{bio.variables}{Character. Column name(s) in \code{colData(se.obj)} representing biological variables for ANOVA.}

\item{uv.variables}{Character. Column name(s) in \code{colData(se.obj)} representing unwanted variation variables.}

\item{nb.bio.clusters}{Numeric. Number of clusters to use for grouping biological variables if clustering is applied.}

\item{bio.clustering.method}{Character. Method used for clustering biological variables (e.g., \code{"kmeans"}, \code{"hierarchical"}).}

\item{nb.uv.clusters}{Numeric. Number of clusters to use for unwanted variation variables if clustering is applied.}

\item{uv.clustering.method}{Character. Method used for clustering unwanted variation variables (e.g., \code{"kmeans"}, \code{"hierarchical"}).}

\item{samples.to.use}{Character or NULL. Vector of sample IDs to include in the analysis. If \code{NULL}, all samples are used.}

\item{apply.log}{Logical. Indicates whether to log-transform the assay data before ANOVA. Default is \code{TRUE}.}

\item{pseudo.count}{Numeric. Pseudo-count added to assay values before log transformation. Default is 1.}

\item{nb.cores}{Numeric. Number of CPU cores to use for parallel computation. Default is 1.}

\item{check.se.obj}{Logical. If \code{TRUE}, validates the \code{SummarizedExperiment} object before analysis. Default is \code{TRUE}.}

\item{remove.na}{Character. Specifies how to handle missing values: \code{"assays"}, \code{"sample.annotation"}, \code{"both"}, or \code{"none"}.
Default is \code{"both"}.}

\item{override.check}{Logical. If \code{TRUE}, recalculates ANOVA even if results already exist in the object metadata. Default is \code{FALSE}.}

\item{save.se.obj}{Logical. If \code{TRUE}, saves F-statistics and p-values in the metadata of the \code{SummarizedExperiment} object;
otherwise returns results as a list. Default is \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE}, displays messages for progress and steps. Default is \code{TRUE}.}
}
\value{
Either a \code{SummarizedExperiment} object containing log2-transformed F-statistics and p-values for ANOVA
or a list of these results.
}
\description{
This function calculates the ANOVA between individual gene expression of the assay(s) in a \code{SummarizedExperiment} object
and a categorical variable as a factor.
}
\details{
ANOVA enables assessment of the effect of a given qualitative variable (factor) on gene expression measurements
across groups labeled by the factor's levels. F-statistics summarize the effects of a qualitative source of unwanted
variation (e.g., batch) on gene expression, with larger F-statistics indicating stronger association. P-values
are assigned using R's \code{aov()} function to quantify associations with variables such as tumor purity or molecular subtypes.
}
\author{
Ramyar Molania
}
