% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGenesVarianceComposition.R
\name{computeGenesVarianceComposition}
\alias{computeGenesVarianceComposition}
\title{Compute gene-level variance composition across variables in RNA-seq.}
\usage{
computeGenesVarianceComposition(
  se.obj,
  assay.names,
  form,
  adjust.data = FALSE,
  adjustment.form = NULL,
  adjustment.method = "lm",
  samples.to.use = "all",
  apply.log = TRUE,
  pseudo.count = 1,
  nb.cores = 1,
  plot.output = TRUE,
  output.name = NULL,
  check.se.obj = TRUE,
  remove.na = "both",
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A \code{SummarizedExperiment} object containing gene expression data and sample annotations.}

\item{assay.names}{Character. A character string or vector of character strings specifying the name(s) of the assay(s)
in the \code{SummarizedExperiment} object to compute gene-level variance composition. By default, set to \code{"all"}, which means all
assays in the object will be selected.}

\item{form}{Formula. A formula specifying the model used for variance partitioning (e.g., \code{~ Batch + Condition + Age}).}

\item{adjust.data}{Logical. If \code{TRUE}, adjusts the data before variance partitioning using the model defined in
\code{adjustment.form}. The default is \code{FALSE}.}

\item{adjustment.form}{Formula. A formula specifying the model to adjust the data (e.g., unwanted variation factors).
Used only if \code{adjust.data = TRUE}.}

\item{adjustment.method}{Character. The method used to adjust for unwanted variation (e.g., \code{"lm"}, \code{"combat"}).}

\item{samples.to.use}{Character or NULL. A vector of sample identifiers to include in the analysis. If \code{NULL},
all samples are used.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data before performing variance
partitioning. The default is set to \code{TRUE}.}

\item{pseudo.count}{Numeric. A numeric value added as a pseudo-count to avoid log-transformation of zeros.
The default is 1.}

\item{nb.cores}{Numeric. Number of CPU cores to use for parallel computation. If \code{NULL}, maximum available cores - 1
will be used.}

\item{plot.output}{Logical. If \code{TRUE}, generates plots summarizing variance composition across variables. The default
is \code{TRUE}.}

\item{output.name}{Character. A label to assign to the results when stored in the metadata of the \code{SummarizedExperiment}
object.}

\item{check.se.obj}{Logical. Indicates whether to validate the \code{SummarizedExperiment} object before running the
analysis. The default is \code{TRUE}.}

\item{remove.na}{Character. Specifies how to handle missing values. Options are \code{"assays"}, \code{"sample.annotation"},
\code{"both"}, or \code{"none"}. The default is \code{"both"}.}

\item{save.se.obj}{Logical. If \code{TRUE}, saves the results (variance components, F-statistics, p-values) in the metadata
of the \code{SummarizedExperiment} object. If \code{FALSE}, returns them as a list. The default is \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE}, displays progress messages during execution. The default is \code{TRUE}.}
}
\value{
Either a \code{SummarizedExperiment} object containing variance partitioning results in the metadata,
or a list of variance components, depending on \code{save.se.obj}.
}
\description{
This function computes the proportion of variance in gene expression explained by different biological
and technical variables in RNA-seq data, using variance partitioning methods. It can adjust for
unwanted variation, apply linear mixed models, and optionally log-transform data.
}
\author{
Ramyar Molania
}
