% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeKbet.R
\name{computeKbet}
\alias{computeKbet}
\title{Compute k-nearest neighbor batch effect test (kBET).}
\usage{
computeKbet(
  se.obj,
  assay.names = "all",
  variable,
  k0 = NULL,
  knn = NULL,
  fast.pca = TRUE,
  nb.pcs = 3,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{character or character vector. A character or a vector of characters of name(s) of the data(assay)
in the SummarizedExperiment object to use for ARI computation. The default is set to \code{all}, meaning all assays in the
object will be selected.}

\item{variable}{character. The name of the column containing a variable in the SummarizedExperiment object.
This variable can represent either a biological or an unwanted factor.}

\item{k0}{number of nearest neighbors to test on (neighborhood size). The default is set to \code{NULL}. Refer to the
\code{kBET} function form the kBET package for more information.}

\item{knn}{an n x k matrix of nearest neighbors for each cell (optional). The default is set to \code{NULL}. Refer to the
\code{kBET} function form the kBET package for more information.}

\item{fast.pca}{Logical. Indicates whether to use principal components computed via fast PCA by the \code{comptePCA} function
The default is set to \code{TRUE}. Note that using fast PCA or standard PCA does not affect silhouette coefficient calculation.
See details for more information.}

\item{nb.pcs}{Numeric. Number of first principal components to use when calculating distances between samples. The default
is set to 3.}

\item{save.se.obj}{Logical. If \code{TRUE}, saves the results in the metadata of the SummarizedExperiment object; otherwise,
returns the result as list. The default is set to \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE}, displays messages for each step of the function.}
}
\value{
A SummarizedExperiment object or a list containing the computed ARI for specified categorical variable.
}
\description{
This function computes k-nearest neighbour batch effect test (kBET) for any given variable using the first PCs of the specified
data(assay) in a SummarizedExperiment object.
}
\details{
The ARI is the corrected-for-chance version of the Rand index. The ARI measures the percentage of matches between
two label lists. We use the ARI to assess the performance of normalization methods in terms of sample subtype
separation and batch mixing. We first calculate PCs and use the first PCs to perform ARI.
}
\references{
Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
\author{
Ramyar Molania
}
