% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLisi.R
\name{computeLisi}
\alias{computeLisi}
\title{Compute local inverse Simpson's index (LISI).}
\usage{
computeLisi(
  se.obj,
  assay.names = "all",
  variable,
  perplexity = 10,
  nn.eps = 0,
  fast.pca = TRUE,
  nb.pcs = 3,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{character or character vector. A character or a vector of characters of name(s) of the data(assay)
in the SummarizedExperiment object to use for ARI computation. The default is set to \code{all}, meaning all assays in the
object will be selected.}

\item{variable}{character. The name of the column containing a variable in the SummarizedExperiment object.
This variable can represent either a biological or an unwanted factor.}

\item{perplexity}{Numeric. The effective number of each cell's neighbors. The default is set to 10.}

\item{nn.eps}{Numeric. Error bound for nearest neighbor search with RANN:nn2(). The default is set to 0.0, implies exact
nearest neighbor search.}

\item{fast.pca}{Logical. Indicates whether to use principal components computed via fast PCA by the \code{comptePCA} function
The default is set to \code{TRUE}. Note that using fast PCA or standard PCA does not affect silhouette coefficient calculation.
See details for more information.}

\item{nb.pcs}{Numeric. Number of first principal components to use when calculating distances between samples. The default
is to 3.}

\item{save.se.obj}{Logical. If \code{TRUE}, saves the results in the metadata of the SummarizedExperiment object; otherwise,
returns the result as list. The default is set to \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE}, displays messages for each step of the function.}
}
\value{
A SummarizedExperiment object or a list containing the computed ARI for specified categorical variable.
}
\description{
This functions computes local inverse Simpson's index (LISI) for any given variable using the first PCs of the specified
data(assay) in a SummarizedExperiment object.
}
\details{
The ARI is the corrected-for-chance version of the Rand index. The ARI measures the percentage of matches between
two label lists. We use the ARI to assess the performance of normalization methods in terms of sample subtype
separation and batch mixing. We first calculate PCs and use the first PCs to perform ARI.
}
\references{
Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
\author{
Ramyar Molania
}
