% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePCA.R
\name{computePCA}
\alias{computePCA}
\title{Performs principal component analysis (PCA)}
\usage{
computePCA(
  se.obj,
  assay.names = "all",
  fast.pca = TRUE,
  nb.pcs = 10,
  center = TRUE,
  scale = FALSE,
  apply.log = TRUE,
  pseudo.count = 1,
  svd.bsparam = bsparam(),
  assess.se.obj = TRUE,
  remove.na = "assays",
  override.check = FALSE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Character. A character string or a vector of character strings for selecting the name(s) of the
assay(s) in the SummarizedExperiment object to compute PCA. The default is set to "all", which indicates all the
assays in the SummarizedExperiment object will be selected.}

\item{fast.pca}{Logical. Indicates whether to calculate a specific number of left singular vectors instead of the
full set of vectors to speed up the process. The default is set to \code{TRUE}.}

\item{nb.pcs}{Numeric. The number of first left singular vectors to be calculated for the fast PCA process.
The default is set to 10. If the \code{fast.pca = FALSE}, no need a value to for \code{nb.pcs}.}

\item{center}{Logical. Indicates whether to center the data before applying SVD. If center is \code{TRUE}, centering is
performed by subtracting the column means of the data from their corresponding columns. The default is set to \code{TRUE}.}

\item{scale}{Logical. Indicates whether to scale the data before applying SVD. If scale is set to \code{TRUE}, scaling is
done by dividing the (centered) columns of the assays by their standard deviations,  if center is \code{TRUE}, and the root
mean square otherwise. The default is set to \code{FALSE}.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data before computing the SVD. The
default is set to \code{TRUE}.}

\item{pseudo.count}{Numeric. A value used as a pseudo-count to be added to all measurements before applying the log
transformation. The default is set to 1.}

\item{svd.bsparam}{A BiocParallelParam object specifying how parallelization should be performed. The default is set
to \code{bsparam()}. See the \code{runSVD()} function from the BiocSingular R package for more details.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object. The default is set to \code{TRUE}.
See the \code{checkSeObj()} function for more details.}

\item{remove.na}{Character. Specifies whether to remove NA or missing values from the data sets (assays). Options are
'assays' and 'none'. The default is set to \code{assays}.}

\item{override.check}{Logical. When set to \code{TRUE}, the function checks whether PCA has already been computed for the
current parameters. If so, the PCA will not be recalculated. The default is set to \code{FALSE}.}

\item{save.se.obj}{Logical. Indicates whether to save the SVD results in the metadata of the SummarizedExperiment object
or to output the results as a list. The default is set to \code{TRUE}. The results can be found:
"se.obj->metadata->metric->AssayName->global.level->PCA$"}

\item{verbose}{Logical. If \code{TRUE}, displays messages for the different steps of the function.}
}
\value{
A SummarizedExperiment object or a list containing the singular value decomposition results and the
percentage variation of each principal component (PC).
}
\description{
This function uses singular value decomposition to perform principal component on the dataset(s) (assay(s) in a
SummarizedExperiment object. The function provides fast singular value decomposition using the BiocSingular R package.
}
\details{
The PCs (in this context also called singular vectors) of the sample × transcript array of log counts are the linear
combinations of the transcript measurements having the largest, second largest, third largest, etc., variation,
standardized to be of unit length and orthogonal to the preceding components. Each will give a single value for
each sample.
}
\author{
Ramyar Molania
}
