% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlateGenes_Variable.R
\name{correlateGenes_Variable}
\alias{correlateGenes_Variable}
\title{is used to find}
\usage{
correlateGenes_Variable(
  se.obj,
  assay,
  variable,
  apply.log = TRUE,
  method = "pearson",
  a = 0.05,
  rho = 0,
  check.top.genes = TRUE,
  top.genes.no = 3
)
}
\arguments{
\item{se.obj}{PCs of the dataset that will be used in the plot}

\item{assay}{All the available assays for the data (i.e. normalizations methods)}

\item{variable}{The regression variable that will be computed to the PCAof the data}

\item{apply.log}{The label of the variable that will be used on the PCA plot}

\item{a}{The number of components of the PCA used to compute the regression}

\item{rho}{The number of components of the PCA used to compute the regression}

\item{check.top.genes}{The number of components of the PCA used to compute the regression}

\item{top.genes.no}{The number of components of the PCA used to compute the regression}
}
\value{
list List containing the association plot and the computed regression
}
\description{
is used to find
}
