% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countNCG.R
\name{countNCG}
\alias{countNCG}
\title{Count and check the overlap the NCG sets in SummarizedExperiment object.}
\usage{
countNCG(
  se.obj,
  ncg.selection,
  ncg.group.name = NULL,
  create.venn.diagram = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{ncg.selection}{Character. A character string indicating which type of NCG should be selected. The options are:
\code{supervised} and \code{unsupervised}.}

\item{ncg.group.name}{Character. A character string indicating which group of NCGs should be selected. Group names are
specified in the functions that identifies NCGs. If set to \code{NULL}, all current NCG group will be selected.}

\item{create.venn.diagram}{Logical. Specifies whether to create a venn diagram using all NCG sets or not. The default
is set to \code{TRUE}.}

\item{verbose}{Logical. Indicates whether to display output messages during function execution. The default is set to
\code{TRUE}.}
}
\description{
This function counts and evaluates the overlap of the NGCs stored in the metadata of the SummarizedExperiment object.
}
\author{
Ramyar Molania
}
