% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPrPsSupervised.R
\name{createPrPsSupervised}
\alias{createPrPsSupervised}
\title{Create PRPS in supervised manner.}
\usage{
createPrPsSupervised(
  se.obj,
  assay.name,
  bio.variables,
  uv.variables,
  apply.other.uv.variables = TRUE,
  other.uv.clustering.method = "kmeans",
  nb.other.uv.clusters = 2,
  samples.to.use = "all",
  min.sample.for.ps = 3,
  bio.clustering.method = "kmeans",
  nb.bio.clusters = 2,
  check.prps.connectedness = TRUE,
  apply.log = TRUE,
  pseudo.count = 1,
  plot.output = TRUE,
  prps.group.name = NULL,
  prps.sets.name = NULL,
  check.se.obj = TRUE,
  remove.na = "both",
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{Character. A character used to specify the assay name within the SummarizedExperiment object.
The selected assay should be the one that will be used for RUV-III-PRPS normalization.}

\item{bio.variables}{Character. A character or character vector representing the label of biological variable(s),
such as cancer subtypes, tumor purity, etc., within the SummarizedExperiment object. This can comprise a vector
containing either categorical, continuous, or a combination of both variables.}

\item{uv.variables}{Character. A character or character vector representing the label of unwanted variable(s),
such as batch effects, library size, etc., within the SummarizedExperiment object. This can comprise a vector containing
either categorical, continuous, or a combination of both variables.}

\item{apply.other.uv.variables}{Logical. Determines whether to include other specified unwanted variables when generating
PRPS sets for individual ones. The default is set to \code{TRUE}.}

\item{other.uv.clustering.method}{Character. A character indicating the clustering method used to group each
continuous unwanted variable. The options include \code{kmeans}, \code{cut}, and \code{quantile}. The default is set to \code{kmeans}.
We refer to the \code{createHomogeneousUVGroups()} function for more details.}

\item{nb.other.uv.clusters}{Numeric. A numeric value to specify the number of clusters/groups for each continuous
unwanted variable. The default is set to 3.}

\item{samples.to.use}{Logical. A logical vector to restrict the analysis to a subset of samples. The default is set to
\code{all}, meaning all samples in the SummarizedExperiment object will be used.}

\item{min.sample.for.ps}{Numeric. Indicates the minimum number of biologically homogeneous samples to be averaged
to create one pseudo-sample. The default is set to 3.}

\item{bio.clustering.method}{Character. A character indicating the clustering method used to group each
continuous biological variable. The options include \code{kmeans}, \code{cut}, and \code{quantile}. The default is set to 'kmeans'.
We refer to the \code{createHomogeneousBioGroups()} function for more details.}

\item{nb.bio.clusters}{Numeric. A numeric value to specify the number of clusters/groups for each continuous biological
variable. The default is set to 3.}

\item{check.prps.connectedness}{Logical. Indicates whether to assess the connectedness between the PRPS sets or not.
The default is set to \code{TRUE}. See the details for more information.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data. The default is set to \code{TRUE}.}

\item{pseudo.count}{Numeric. A numeric value as a pseudo count to be added to all measurements before log transformation.
The default is set to 1.}

\item{plot.output}{Logical. Indicates whether to generate the PRPS map plot for individual sources of unwanted variation.
The default is set to \code{TRUE}.}

\item{prps.group.name}{Character. A character specifying a name for the PRPS data sets created for a specific group.
The default is set to \code{NULL}.}

\item{prps.sets.name}{Character. A character to specify the name of all PRPS sets that will be created for all
specified source(s) of unwanted variation. The default is set to \code{NULL}. If not specified, the function creates a
name based on \code{paste0('prps_', uv.variable)}.}

\item{check.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object. If \code{TRUE}, the \code{checkSeObj()}
function will be applied inside the function. The default is set to \code{TRUE}.}

\item{remove.na}{Character. A character indicating whether to remove NA or missing values from either the 'assays',
the \code{sample.annotation}, \code{both}, or \code{none}. If \code{assays} is selected, the genes that contain NA or missing values will
be excluded. If 'sample.annotation' is selected, the samples that contain NA or missing values for any \code{bio.variables}
and \code{uv.variables} will be excluded. The default is set to \code{both}.}

\item{save.se.obj}{Logical. Indicates whether to save the results in the metadata of the SummarizedExperiment object
or to output the result as a list. The default is set to \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE}, shows the messages of different steps of the function.}
}
\value{
Either a SummarizedExperiment object that contains all the PRPS data and PRPS map plots in the metadata, or a
list that contains all the results.
}
\description{
This function creates different PRPS sets when all sources of unwanted and biological variation are known.
}
\details{
We will create distinct group of pseudo-replicates for each source of unwanted variation defined in the \code{uv.variables}
argument. For example to correct for batch effect if defined in the \code{uv.variables} argument, several group of
pseudo-samples will be created by averaging the samples of the same biological groups defined in 'bio.variables' in
each batch. Then those pseudo-samples will be defined as pseudo-replicates which constitutes a PRPS set. For example
to correct for library size if defined in the 'uv.variables' argument, several group of pseudo-samples will be created
by averaging the top and bottom-ranked samples by library size of the same biological subtype in each batch. Then those
pseudo-samples will be defined as pseudo-replicates which constitutes a PRPS set. Similarly to correct for purity if
defined in the uv.variables' argument, several group of pseudo-samples will be created by averaging the top and
bottom-ranked samples by purity of the same biological subtype in each batch. Then those pseudo-samples will be defined
as pseudo-replicates which constitutes a PRPS set.
}
\references{
Molania R., ..., Speed, T. P., A new normalization for Nanostring nCounter gene expression data, Nucleic Acids Research,
2019.
Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
\author{
Ramyar Molania
}
