% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPrPsUnsupervisedByCca.R
\name{createPrPsUnSupervisedByCca}
\alias{createPrPsUnSupervisedByCca}
\title{Creates PRPS sets using mutual nearest neighbors in RNA-seq data.}
\usage{
createPrPsUnSupervisedByCca(
  se.obj,
  assay.name,
  main.uv.variable,
  reference.group = NULL,
  other.uv.variables = NULL,
  coordinates.to.use = "both",
  nb.cca = 2,
  nb.pcs = 5,
  samples.to.use = "all",
  min.sample.for.ps = 3,
  select.extreme.groups = FALSE,
  filter.prps.sets = TRUE,
  max.prps.sets = 3,
  min.batches.to.cover = "all",
  cover.all.batches = FALSE,
  check.prps.connectedness = FALSE,
  hvg = NULL,
  scale.cca = TRUE,
  apply.ruviii.norm = TRUE,
  use.ruviii.norm.for.mnn = TRUE,
  ncg = NULL,
  k = 2,
  nb.mnn = 3,
  min.ps = 10,
  min.nb.for.mnn = 1,
  similarity.approach = "euclidean",
  data.for.similarity = "ruv",
  clustering.method = "cut",
  nb.clusters = 3,
  other.uv.clustering.method = "kmeans",
  nb.other.uv.clusters = 2,
  nb.batches.to.cover = 2,
  normalization = "CPM",
  cosine.norm = FALSE,
  regress.out.variables = NULL,
  regress.out.rle.med = FALSE,
  apply.log = TRUE,
  apply.log.for.prps = TRUE,
  pseudo.count = 1,
  assess.variables.association = TRUE,
  create.prps.map = FALSE,
  plot.output = TRUE,
  mnn.bpparam = SerialParam(),
  mnn.nbparam = KmknnParam(),
  residop.fun = "r2",
  check.se.obj = TRUE,
  remove.na = "both",
  cca.set.name = NULL,
  prps.group.name = NULL,
  prps.sets.name = NULL,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{Character. A character indicating the name of the data (assay) in the SummarizedExperiment object.
This data will be used to create PRPS data for RUV-III normalization. This data must be the one that will be
used for the RUV-III normalization.}

\item{main.uv.variable}{Character. Indicates the name of a column in the sample annotation of the SummarizedExperiment
object. The \code{uv.variable} can be either categorical or continuous. If \code{uv.variable} is a continuous variable, this will
be divided into \code{nb.clusters} groups using the \code{clustering.method}.}

\item{reference.group}{Character. A character string specifying which group in the \code{main.uv.variable} should be used
as the reference group when constructing PRPS. Default is set to \code{NULL}, meaning no explicit reference group is used.}

\item{other.uv.variables}{Character. A character or character vector representing the name(s) of the columns of
unwanted variable(s) within the sample annotation (colData) of the SummarizedExperiment object. These can be categorical,
continuous, or a combination. These variables will be considered when generating PRPS sets for the \code{main.uv.variable}
to help avoid potential contamination. The default is set to \code{NULL}}

\item{coordinates.to.use}{Character. Indicates which coordinates (e.g., PCA or CCA embeddings) should be used to
identify mutual nearest neighbors. Default is set to \code{NULL}, meaning the raw assay data is used.}

\item{nb.cca}{Numeric. Number of canonical correlation components to use for constructing PRPS. Default is set to 20.}

\item{nb.pcs}{Numeric. Number of principal components to use for constructing PRPS. Default is set to 20.}

\item{samples.to.use}{Character. A vector of sample names to restrict the analysis to. Default is \code{NULL}, meaning
all samples in the SummarizedExperiment object are used.}

\item{min.sample.for.ps}{Numeric. Minimum number of samples required for pseudo-replicate creation. The default is set
to 3.}

\item{select.extreme.groups}{Logical. Indicates whether to select only the extreme groups e.g., highest and lowest
clusters, when the \code{uv.variable} is continuous. Default is set to \code{TRUE}. This will increase the variation between
PR sets in order to better capture the unwanted variation.}

\item{filter.prps.sets}{Logical. If \code{TRUE}, the number of PRPS sets across each pair of batches will be filtered if
they are higher than the \code{max.prps.sets} value. The default is set to \code{TRUE}. A high number of PRPS sets will increase
the computational time for the RUV-III normalization.}

\item{max.prps.sets}{Numeric. A numeric value specifying the maximum number for PRPS sets across each pair of batches.
The default is set to 10.}

\item{min.batches.to.cover}{Numeric. Minimum number of batches that must be covered by PRPS set. The default is set to
\code{all}, indicating all possible batch must have enough samples to create PRPS, otherwise the function gives error.}

\item{cover.all.batches}{Logical. Indicates whether PRPS must be created such that all batches are covered. If \code{TRUE},
the function ensures coverage across all batches; otherwise, partial coverage is allowed. Default is set to \code{TRUE}.}

\item{check.prps.connectedness}{Logical. Indicates whether to assess the \code{connectedness} between the PRPS sets across
all batches. Default is set to \code{TRUE}, indicating if there is not connections between all PRPS sets across all batches,
the function will give error. We refer to the checkPRPSconnectedness() function for more details.}

\item{hvg}{Vector. A logical vector or a vector of the names (feature ids) of the highly variable genes. These genes
will be used to prepare the input data for knn and mnn analysis. The default is set to \code{NULL}, this means all genes
will be used.}

\item{scale.cca}{Logical. Indicates whether to scale the CCA components before constructing PRPS. Default is set to \code{TRUE}.}

\item{apply.ruviii.norm}{Logical. Indicates whether to apply RUV-III normalization on the constructed PRPS before
downstream analysis. Default is set to \code{FALSE}.}

\item{use.ruviii.norm.for.mnn}{Logical. Indicates whether RUV-III normalized data should be used to identify mutual
nearest neighbors. Default is set to \code{FALSE}.}

\item{ncg}{Character. A vector of negative control genes to be used in the RUV-III normalization when
\code{apply.ruviii.norm = TRUE}. Default is set to \code{NULL}.}

\item{k}{Numeric. The number of nearest neighbors to use when finding mutual nearest neighbors. Default is set to 20.}

\item{nb.mnn}{Numeric. A numeric value specifying the maximum number of mutual nearest neighbors to compute. The
default is set to 1.}

\item{min.ps}{Numeric. Minimum number of pseudo-samples required for each PRPS set to be considered valid. Default is set to 2.}

\item{min.nb.for.mnn}{Numeric. Minimum number of neighbors required to define a mutual nearest neighbor pair. Default is set to 5.}

\item{similarity.approach}{Character. A character string specifying how similarity between samples is measured
(e.g., \code{cosine}, \code{correlation}, or \code{euclidean}). Default is set to \code{cosine}.}

\item{data.for.similarity}{Character. Specifies which type of data should be used for computing similarity
(e.g., \code{raw}, \code{normalized}, or \code{reduced}). Default is set to \code{normalized}.}

\item{clustering.method}{Character. A character indicating the choice of clustering method for grouping the
\code{uv.variable} if a continuous variable is provided. Options include \code{kmeans}, \code{cut}, and \code{quantile}. The default is set
to \code{kmeans}.}

\item{nb.clusters}{Numeric. A numeric value indicating how many clusters should be found if the \code{uv.variable} is a
continuous variable. The default is set to 3.}

\item{other.uv.clustering.method}{Character. A character indicating which clustering method should be used to
group each continuous unwanted variable, if specified in \code{other.uv.variables}. Options include \code{kmeans}, \code{cut},
and \code{quantile}. The default is set to \code{kmeans}. See createHomogeneousUVGroups() for more details.}

\item{nb.other.uv.clusters}{Numeric. A numeric value to specify the number of clusters/groups for each continuous
unwanted variable specified in the \code{other.uv.variables}. The default is set to 3.}

\item{nb.batches.to.cover}{Numeric. A numeric value specifying the minimum number of batches that each PRPS set must
span to be considered valid. Default is set to 2.}

\item{normalization}{Character. A character that indicates which normalization method should be applied on the
data before finding the knn. Options are: \code{CPM}, \code{TMM}, \code{upper}, \code{median}, \code{full}, and \code{VST}. The default is set to
\code{cpm}.
If set to \code{NULL}, no normalization will be applied. See the applyOtherNormalizations() function for more details.}

\item{cosine.norm}{Logical. Indicates whether cosine normalization should be applied before finding MNN. Default
is set to \code{TRUE}.}

\item{regress.out.variables}{Character. A character or a vector of character that indicate the column name(s) in the
sample annotation in the SummarizedExperiment object. These variables will be regressed out from the data before
finding MNN. The default is set to \code{NULL}, indicating that regression will not be applied.}

\item{regress.out.rle.med}{Logical. Indicates whether to regress out the relative log expression (RLE) median
from the data before finding MNN. Default is set to \code{FALSE}.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data or not for down-stream analysis.
The default is set to \code{TRUE}.}

\item{apply.log.for.prps}{Logical. Indicates whether log-transformation should be applied only for the construction of
PRPS sets, regardless of global log-transformation. Default is set to \code{FALSE}.}

\item{pseudo.count}{Numeric. A positive numeric value as a pseudo count to be added to all measurements of the specified
assay(data) before applying log transformation to avoid -Inf for measurements that are equal to 0. The default is set
to 1.}

\item{assess.variables.association}{Logical. Indicates whether to assess the association between constructed PRPS and
biological/unwanted variables. Default is set to \code{TRUE}.}

\item{create.prps.map}{Logical. Indicates whether to generate a visualization (map) of the constructed PRPS sets across
batches and unwanted variables. Default is set to \code{FALSE}.}

\item{plot.output}{Logical. If \code{TRUE}, the function plots the distribution of MNN across the batches and PRPS sets
across the \code{main.uv.variable}.}

\item{mnn.bpparam}{Character. A BiocParallelParam object specifying how palatalization should be performed to find MNN.
The default is set to SerialParam(). We refer to the \code{findMutualNN()} function from the \strong{BiocNeighbors} R package.}

\item{mnn.nbparam}{Character. A BiocParallelParam object specifying how parallelization should be performed to find MNN.
The default is KmknnParam(). We refer to the \code{findMutualNN()} function from the \code{BiocNeighbors} R package.}

\item{residop.fun}{Character. A character indicating which function to use to calculate residuals. The options are
\code{c1}, \code{c2}, \code{lqr}, \code{r1} and \code{r2}. The default is set to \code{r2}.}

\item{check.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object or not. The default is set
to \code{TRUE}. See the checkSeObj() function for more details.}

\item{remove.na}{Character. To remove NA or missing values from the assay (data) or not. The options are \code{assays} and
\code{none}. The default is set to \code{assays}, so all the NA or missing values from the assay(s) will be removed before computing
performing any down-stream analysis. See the \code{checkSeObj()} function for more details.}

\item{cca.set.name}{Character. A character string specifying the name of the CCA set used for PRPS construction.
Default is set to \code{NULL}, meaning a name will be automatically generated.}

\item{prps.group.name}{Character. A character specifying the name of the prps.group.name to which the current KNN belong.
If set to \code{NULL}, the function will automatically assign a name using  \code{main.uv.variable}.}

\item{prps.sets.name}{Character. A character specifying the name of the output file to be saved in the metadata
of the SummarizedExperiment object. If set to \code{NULL}, the function will select a name based on
\code{paste0(uv.variable, '|', assay.name)}.}

\item{save.se.obj}{Logical. Indicates whether to save the KNN results in the metadata of the SummarizedExperiment object
or to output the result as a list. The default is set to \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE}, shows the messages of different steps of the function.}
}
\value{
The SummarizedExperiment object containing PRPS data and plot results in the metadata, or a list of
these results.
}
\description{
This function uses mutual nearest neighbors approach to create PRPS in the RNA-seq data. This function can be used in
situation where the biological variation are entirely unknown.
}
\author{
Ramyar Molania
}
