% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCMS.R
\name{estimateCMS}
\alias{estimateCMS}
\title{Estimates the CMS subtypes of colorectal cancer RNA-seq data.}
\usage{
estimateCMS(
  se.obj,
  assay.names = "all",
  raw.count.assay.name = NULL,
  subtype = "CMS",
  tissue.type = NULL,
  groups = NULL,
  templates = CMScaller::templates.CMS,
  row.names = "symbol",
  count.data = FALSE,
  nb.perm = 1000,
  fdr = 0.05,
  generate.plot = FALSE,
  nb.cores = 1,
  apply.log = TRUE,
  pseudo.count = 1,
  out.put.name = NULL,
  check.se.obj = TRUE,
  remove.na = "none",
  seed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{description}

\item{assay.names}{description}

\item{raw.count.assay.name}{description}

\item{subtype}{description}

\item{tissue.type}{description}

\item{groups}{description}

\item{templates}{description}

\item{row.names}{description}

\item{count.data}{description}

\item{nb.perm}{description}

\item{fdr}{description}

\item{generate.plot}{description}

\item{nb.cores}{description}

\item{apply.log}{description}

\item{pseudo.count}{description}

\item{out.put.name}{description}

\item{check.se.obj}{description}

\item{remove.na}{description}

\item{seed}{description}

\item{verbose}{description}
}
\description{
This function employees the \code{CMScaller} function from the \code{CMScaller} R package to estimates the CMS subtypes of
colorectal cancer RNA-seq data
}
\author{
Ramyar Molania
}
