% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePAM50.R
\name{estimatePAM50}
\alias{estimatePAM50}
\title{Estimate PAM50 subtypes}
\usage{
estimatePAM50(
  se.obj,
  assay.names = "all",
  entrez.gene.id,
  probe,
  tissue.type = NULL,
  groups = NULL,
  apply.log = TRUE,
  pseudo.count = 1,
  normalization = "CPM",
  regress.out.variables = NULL,
  regress.out.rle.med = FALSE,
  out.put.name = NULL,
  check.se.obj = TRUE,
  remove.na = "none",
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Character or character vector. The name(s) of the assay(s) in the \code{SummarizedExperiment} object.
The default is set to 'all'.}

\item{entrez.gene.id}{description}

\item{probe}{description}

\item{tissue.type}{description}

\item{groups}{description}

\item{apply.log}{description}

\item{pseudo.count}{description}

\item{normalization}{description}

\item{regress.out.variables}{description}

\item{regress.out.rle.med}{description}

\item{out.put.name}{description}

\item{check.se.obj}{description}

\item{remove.na}{description}

\item{verbose}{description}
}
\description{
Estimate PAM50 subtypes
}
\author{
Ramyar Molania
}
