% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findHVG.R
\name{findHVG}
\alias{findHVG}
\title{Find highly variable genes on RNA-seq data.}
\usage{
findHVG(
  se.obj,
  assay.name,
  approach = "vst",
  nb.hvg = 0.1,
  uv.variables = NULL,
  bio.variables = NULL,
  group.name = NULL,
  hvg.selection = "intersect",
  form = NULL,
  adjust.data = FALSE,
  nb.bio.pcs = 3,
  bio.percentile = 0.7,
  span = 0.3,
  clustering.method = "kmeans",
  nb.clusters = 3,
  samples.to.use = "all",
  normalization = NULL,
  regress.out.variables = NULL,
  regress.out.rle.med = FALSE,
  center = TRUE,
  scale = FALSE,
  svd.bsparam = bsparam(),
  apply.log = TRUE,
  pseudo.count = 1,
  nb.cores = 1,
  check.se.obj = TRUE,
  remove.na = "none",
  hvg.group.name = NULL,
  hvg.set.name = NULL,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{description}

\item{assay.name}{description}

\item{approach}{description}

\item{nb.hvg}{description}

\item{uv.variables}{description}

\item{bio.variables}{description}

\item{group.name}{description}

\item{hvg.selection}{TTT}

\item{form}{description}

\item{adjust.data}{description}

\item{nb.bio.pcs}{description}

\item{bio.percentile}{TTT}

\item{span}{description}

\item{clustering.method}{description}

\item{nb.clusters}{description}

\item{samples.to.use}{description}

\item{normalization}{description}

\item{regress.out.variables}{description}

\item{regress.out.rle.med}{description}

\item{center}{TTT}

\item{scale}{TTT}

\item{svd.bsparam}{TTT}

\item{apply.log}{description}

\item{pseudo.count}{description}

\item{nb.cores}{TTT}

\item{check.se.obj}{TTT}

\item{remove.na}{description}

\item{hvg.group.name}{description}

\item{hvg.set.name}{description}

\item{save.se.obj}{description}

\item{verbose}{description}
}
\description{
Find highly variable genes on RNA-seq data.
}
\author{
Ramyar Molania
}
