% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNcgPerBiologyPerBatch.R
\name{findNcgPerBiologyPerBatch}
\alias{findNcgPerBiologyPerBatch}
\title{Finds NCGs using ANOVA and correlation per sample groups.}
\usage{
findNcgPerBiologyPerBatch(
  se.obj,
  assay.name,
  bio.variables,
  uv.variables,
  ncg.selection.method = "non.overlap",
  nb.ncg = 0.1,
  top.rank.bio.genes = 0.5,
  top.rank.uv.genes = 0.5,
  bio.percentile = 0.2,
  uv.percentile = 0.2,
  grid.group = "uv",
  grid.direction = "decrease",
  grid.nb = 20,
  create.ncg.rank.plot = FALSE,
  plot.ncg.rank = FALSE,
  min.sample.for.aov = 3,
  min.sample.for.correlation = 10,
  regress.out.bio.variables = NULL,
  regress.out.uv.variables = NULL,
  bio.groups = NULL,
  bio.clustering.method = "kmeans",
  nb.bio.clusters = 3,
  uv.groups = NULL,
  uv.clustering.method = "kmeans",
  nb.uv.clusters = 3,
  normalization = "CPM",
  apply.log = TRUE,
  pseudo.count = 1,
  corr.method = "spearman",
  a = 0.05,
  rho = 0,
  anova.method = "aov",
  assess.ncg = TRUE,
  variables.to.assess.ncg = NULL,
  nb.pcs = 5,
  center = TRUE,
  scale = FALSE,
  plot.ncg.assessment = TRUE,
  check.se.obj = TRUE,
  remove.na = "none",
  ncg.group.name = NULL,
  ncg.set.name = NULL,
  save.imf = FALSE,
  imf.name = NULL,
  use.imf = FALSE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{Character. A character that indicates the name of an data (assay) in the SummarizedExperiment object.
The selected assay should be the one that will be used for the RUV-III-PRPS normalization.}

\item{bio.variables}{Character. A character string or vector of strings indicating the column name(s) of the biological
variable(s) in the SummarizedExperiment object. These variable can be categorical or continuous or a combination. This
argument cannot be \code{NULL}.}

\item{uv.variables}{Character. A character string or vector of strings indicating the column name(s) of the unwanted
variable(s) in the SummarizedExperiment object. These variable can be categorical or continuous or a combination.This
argument cannot be \code{NULL}.}

\item{ncg.selection.method}{Character. A character that indicates how to summarize different statistics and select a
set of genes as negative control genes. The options are: \code{prod}, \code{average}, \code{sum}, \code{non.overlap}, \code{auto}, and \code{quantile}.
The default is set to \code{non.overlap}. For more information, refer to the details of the function.}

\item{nb.ncg}{Numeric. A numeric value that specifies the number of genes to be chosen as negative control genes (NCG)
when the \code{ncg.selection.method} parameter is set to \code{auto}. This value, \code{nb.ncg}, corresponds to a fraction of the total
genes in the SummarizedExperiment object. The default is set to 0.1.}

\item{top.rank.bio.genes}{Numeric. A numeric value that indicates the percentage of top-ranked genes that are highly
affected by biological variation. This is required to be specified when the \code{ncg.selection.method} is either \code{auto}
or \code{non.overlap}. The default is set to 0.2.}

\item{top.rank.uv.genes}{Numeric. A numeric value that indicates the percentage of top-ranked genes that are highly
affected by unwanted variables. This is required to be specified when the \code{ncg.selection.method} is either \code{auto} or
\code{non.overlap}. The default is set to 0.2.}

\item{bio.percentile}{Numeric. A numeric value that specifies the percentile cut-off to select genes that are highly
affected by biological variation. This is required to be specified when the \code{ncg.selection.method} is set to \code{quantile}.
The default is set to 0.8.}

\item{uv.percentile}{Numeric. A numeric value that specifies the percentile cut-off to select genes that are highly
affected by unwanted variation. This is required to be specified when the \code{ncg.selection.method} is set to \code{quantile}.
The default is set to 0.8.}

\item{grid.group}{Character. A character that indicates whether the grid search should be performed on biological,
unwanted, or both factors when the \code{ncg.selection.method} is set to \code{auto}. The options are \code{bio}, \code{uv}, or \code{both}.
The default is set to \code{uv}. For more details, refer to the function documentation.}

\item{grid.direction}{Character. A character that indicates whether the grid search should be performed in decreasing
or increasing order when the \code{ncg.selection.method} is set to \code{auto}. The options are: \code{increase} and \code{decrease}. The
default is set to \code{decrease}.}

\item{grid.nb}{Numeric. A numeric value that indicates the number of genes for grid search when the \code{ncg.selection.method}
is set to \code{auto}. In the \code{auto} approach, the grid search starts with the initial \code{top.rank.uv.genes} and
\code{top.rank.bio.genes} values and adds or drops the \code{grid.nb} in each loop to find \code{nb.ncg} of genes as negative control
genes. The default is set to 20.}

\item{create.ncg.rank.plot}{Logical. Indicates whether to generate a heatmap that shows the rank of the all genes
with respect to their biological and unwanted variation effects. The default is set to \code{FALSE}.}

\item{plot.ncg.rank}{Logical. Indicates whether to plot a heatmap that shows the rank of the all genes
with respect to their biological and unwanted variation effects, while function is running. The default is set to \code{FALSE}.}

\item{min.sample.for.aov}{Numeric. A numeric value that indicates the minimum number of samples that are required to
perform ANOVA analyses between continuous sources of variation (biological and unwanted variation) with individual
gene expression. The default is set to 3. The minimum value is 3.}

\item{min.sample.for.correlation}{Numeric. A numeric value that indicates the minimum number of samples that are required
to perform correlation analyses between continuous sources of variation (biological and unwanted variation) with
individual gene expression. The default is set to 10. The minimum value can be 3.}

\item{regress.out.bio.variables}{Character. A character string or vector of strings that indicate the column names of
biological variables in the SummarizedExperiment object that will be regressed out from the data before performing
correlation and ANOVA. Regressing out biological variables might help better identify genes that are highly affected
by unwanted variation. The default is set to \code{NULL}.}

\item{regress.out.uv.variables}{Character. A character string or vector of strings that indicate the column names of
unwanted variables in the SummarizedExperiment object that will be regressed out from the data before performing
correlation and ANOVA. Regressing out unwanted variables might help better identify genes that are highly affected by
biological variation. The default is set to \code{NULL}.}

\item{bio.groups}{A character vector or a symbol. Indicates the column names that contain biological variables in the
SummarizedExperiment object. If not NULL, bio.groups' will be used to group samples into different homogeneous biological groups.}

\item{bio.clustering.method}{A character string. Indicates which clustering method should be used to group continuous
sources of biological variation if any are provided. The default is 'kmeans' clustering.}

\item{nb.bio.clusters}{Numeric. Indicates the number of clusters for each continuous source of biological variation.
The default is set to 3.}

\item{uv.groups}{A character vector or a symbol. Indicates the column names that contain unwanted variation variables in the
SummarizedExperiment object. If not NULL, \code{uv.groups} will be used to group samples into different homogeneous unwanted
variation groups.}

\item{uv.clustering.method}{A character string. Indicates which clustering method should be used to group continuous
sources of unwanted variation. The default is set to 'kmeans' clustering.}

\item{nb.uv.clusters}{Numeric. Indicates the number of clusters for each continuous source of unwanted variation (UV).
By default, it is set to 2.}

\item{normalization}{Character. A character that indicates which normalization method should be use to mitigate the
variation in library size before finding genes that are highly affected by biological variation. The options are :
\code{CPM}, \code{TMM}, \code{VST}, \code{upper}, \code{full} and \code{medium}. The default is set to  \code{CPM}. Refer to the \code{applyOtherNormalization()}
function for more details.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data before any statistical analyses.
The default is set to \code{TRUE}.}

\item{pseudo.count}{Numeric. A numeric value to be added as a pseudo count to all measurements before applying log
transformation. The default is set to 1.}

\item{corr.method}{Character. A character that indicates which correlation methods should be used for the correlation
analyses. The options are \code{pearson} or \code{spearman}. The default is set to \code{spearman}.}

\item{a}{Numeric. The significance level used for the confidence intervals in the correlation; by default, it is set
to 0.05. Refer to the function \code{correls} from the \strong{Rfast} R package for more details.}

\item{rho}{Numeric. The value of the hypothesized correlation to be used in the hypothesis testing. The default is
set to 0. Refer to the function \code{correls} from the \strong{Rfast} R package for more details.}

\item{anova.method}{Character. A character that  indicates which ANOVA method to use. The options are \code{aov} or \code{welch}.
The default is se to \code{aov}. Refer to the function \code{row_oneway_equalvar()} or \code{row_oneway_welch()} from the
\strong{matrixTests} R package for more details.}

\item{assess.ncg}{Logical. Indicates whether to assess the performance of selected genes as negative control or not.
This analysis involves principal component analysis on the selected genes, followed by exploration of the R^2 or vector
correlation between the first \code{nb.pcs} principal components and the biological and unwanted variables. The default is
set to \code{TRUE}.}

\item{variables.to.assess.ncg}{Character. A character string or vector of strings indicating the column names in sample
annotation of of the  SummarizedExperiment object that contain variables whose association with the selected genes as
NCG needs to be evaluated. The default is set to \code{NULL}. This means all the variables specified in the \code{bio.variables}
and \code{uv.variables} will be assessed.}

\item{nb.pcs}{Numeric. A numeric value that indicates the number of the first principal components of selected negative
control genes to be used to assess their performance. The default is set to 10.}

\item{center}{Logical. Indicates whether to scale the data before applying SVD. If \code{TRUE}, centering is done by subtracting
the column means of the assay from their corresponding columns. The default is set to \code{TRUE}.}

\item{scale}{Logical. Indicates whether to scale the data before applying SVD. If \code{TRUE}, scaling is done by dividing the
(centered) columns of the assays by their standard deviations if centering is \code{TRUE}, and by the root mean square otherwise.
The default is set to \code{FALSE}.}

\item{plot.ncg.assessment}{Logical. Indicates whether to plot the output of the NCG assessment while function is running
. The default is set to \code{TRUE}.}

\item{check.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object before any analysis. If \code{TRUE},
the function \code{checkSeObj()} will be used. The default is set to \code{TRUE}.}

\item{remove.na}{Character set. Indicates whether to remove NA or missing values from the SummarizedExperiment object
The options are: \code{assays}, the \code{sample.annotation}, \code{both}, or \code{none}. If \code{assays} is selected, genes containing NA or
missing values will be excluded. If \code{sample.annotation} is selected, the samples containing NA or missing values for
any \code{bio.variables} or \code{uv.variables} will be excluded. The default is set to \code{none}.}

\item{ncg.group.name}{Character. A character to be used as name of the group of NCG. The default is set to \code{NULL}, then
the function create a names as following: \code{paste0('ncg|unsupervised')}. We refer to the details of the function for
more details.}

\item{ncg.set.name}{Character. A character to be used as name of the NCG set based on current variables and parameters
The default is set to \code{NULL}, then the function create a names as following:
\code{paste0(sum(ncg.selected),'|',paste0(bio.variables, collapse = '&'),'|',paste0(uv.variables, collapse = '&'),'|AnoCorrAs:', ncg.selection.method,'|',assay.name)}.We refer to the details of the function for more details.}

\item{save.imf}{Logical. Indicates whether to save the intermediate file. If \code{TRUE}, the function saves the results
of the statistical analyses in the metadata of the SummarizedExperiment object. If users want to change the parameters
including \code{nb.ncg}, \code{ncg.selection.method}, \code{top.rank.bio.genes}, and \code{top.rank.uv.genes}, the analyses will not be
re-calculated. The default is set to \code{FALSE}.}

\item{imf.name}{Character string. A name to save the intermediate file. If \code{NULL}, the function generates a name.}

\item{use.imf}{Logical. Indicates whether to use the intermediate file. The default is set to \code{FALSE}.}

\item{save.se.obj}{Logical. Indicates whether to save the result of the function in the metadata of the
SummarizedExperiment object or output the result. The default is \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE}, shows messages of different steps of the function.}
}
\value{
Either the SummarizedExperiment object containing a set of negative control genes in the metadata or a
logical vector of the selected negative control genes.
}
\description{
This function employs correlation and ANOVA analyses within groups of samples that are homogeneous with respect
to biological or unwanted variation. Correlation analysis is utilized to identify genes highly affected by continuous
sources of variation, while ANOVA is used to identify genes affected by categorical sources of variation.
}
\details{
The function selects genes as negative control genes (NCG) based on high correlation coefficients and F-statistics for
unwanted sources of variation, and low correlation coefficients and F-statistics for biological sources of variation.
Various approaches are employed for the final gene selection; please refer to the details for more information.
The function uses 5 ways to summarize two gene-level F-statistics obtained for the biological and unwanted variation
. The function uses either the values or the ranks of F-statistics for NCGs selection. The function ranks the
negative of F-statistics values for unwanted variation. The lower the ranks, the greater the impact of unwanted
variation on genes. The function ranks the F-statistics for biological variation. The higher the ranks, the greater
the impact of biological variation on genes. The options are 'prod', sum', 'average', 'auto' or 'non.overlap' and
'quantile'.

If 'prod', 'sum' and 'average' is set:

The product, sum or average of ranks of F-statistics is calculated. Then, the function selects 'nb.ncg' 'numbers of
genes as negative control genes that have the lowest ranks.

If 'non.overlap' is selected:
\enumerate{
\item The function selects the top ‘top.rank.bio.genes’ genes that have the highest ranks of F-statistics
for biological variation.
\item The function selects the top ‘top.rank.uv.genes’ genes that have the lowest ranks of F-statistics for
unwanted variation.
\item The function excludes all genes obtained in 2 from the ones obtained 1. This will be a set of genes as
negative control genes.
}

If 'auto' is selected:
\enumerate{
\item The function selects the top ‘top.rank.bio.genes’ genes that have the highest ranks of F-statistics for
biological variation.
\item  The function selects the top ‘top.rank.uv.genes’ genes that have the lowest ranks of F-statistics  for
unwanted variation.
\item The function excludes all genes obtained in 2 from the ones obtained 1.
\item If the number of selected genes is larger or smaller than the specified ‘nb.ncg’, the function applies an
auto search to find approximate ‘nb.ncg’ of genes as negative control genes as follow. The auto search will either
decrease or increase the values of either ‘top.rank.bio.genes’ or ‘top.rank.uv.genes’ or both till to find
approximate ‘nb.ncg’ of genes as negative control genes.
}
If 'quantile' is selected:
\enumerate{
\item The function selects the ‘bio.percentile’ percentile of F-statistics for biological variation. Then, selects
all the genes that have F-statistics larger the calculated percentile.
\item The function selects the ‘uv.percentile’ percentile of F-statistics for unwanted variation. Then, selects
all the genes that have F-statistics larger the calculated percentile.
\item The function excludes all genes obtained in 2 from the ones obtained 1.
}

The data pre-processing for the functions findNcgAcrossSamples() and findNcgPerBiologyPerBatch().
The input data for correlation and ANOVA analyses can be one or a combination of the following options:
\enumerate{
\item item $log_2$ (raw data + pseudo-count).
\item Apply some library size normalizations (if the input data is gene count expression).
\item This will facilitate finding genes that are highly affected by biological variation. item Regress out
biological variation.
\item This will facilitate finding genes that are highly affected by unwanted variation.
item Regress out unwanted variation.
}
}
\author{
Ramyar Molania
}
