% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNcgUnSupervised.R
\name{findNcgsUnSupervised}
\alias{findNcgsUnSupervised}
\title{Finds negative control genes (NCG) using an unsupervised approach.}
\usage{
findNcgsUnSupervised(
  se.obj,
  assay.name,
  uv.variables,
  clustering.method = "kmeans",
  nb.clusters = 3,
  nb.ncg = 0.1,
  hvg.method = "var",
  ncg.selection.method = "non.overlap",
  top.rank.bio.genes = 0.5,
  top.rank.uv.genes = 0.5,
  bio.percentile = 0.2,
  uv.percentile = 0.2,
  grid.group = "uv",
  grid.direction = "decrease",
  grid.nb = 20,
  normalization = "CPM",
  regress.out.variables = NULL,
  min.sample.for.mad = 3,
  min.sample.for.var = 15,
  min.sample.for.aov = 3,
  min.sample.for.correlation = 10,
  corr.method = "spearman",
  a = 0.05,
  rho = 0,
  anova.method = "aov",
  filter.ncgs = FALSE,
  common.hk = "cancer",
  hk.group = "micorarray.hk.genes",
  assess.ncg = TRUE,
  variables.to.assess.ncg = NULL,
  nb.pcs = 5,
  center = TRUE,
  scale = FALSE,
  apply.log = TRUE,
  pseudo.count = 1,
  assess.se.obj = TRUE,
  remove.na = "both",
  output.name = NULL,
  ncg.group = NULL,
  plot.output = TRUE,
  output.plot = "assessment",
  use.imf = FALSE,
  save.imf = FALSE,
  imf.name = NULL,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A \code{SummarizedExperiment} object.}

\item{assay.name}{Character. The name of the assay in the \code{SummarizedExperiment} object to be used to find NCGs.
This must be the same data used for \code{RUV-III} normalization.}

\item{uv.variables}{Character. A character vector indicating the name(s) of the columns in the \code{SummarizedExperiment} object
that contain unwanted variable(s). If all unwanted variation is unknown, use the \code{identifyUnknownUV()} function
to estimate them first, then run \code{findNcgsUnSupervised()}.}

\item{clustering.method}{Character. The clustering method used to group continuous unwanted variation. Options are
\code{kmeans}, \code{cut}, and \code{quantile}. The default is set to \code{kmeans}.}

\item{nb.clusters}{Numeric. Number of clusters for grouping continuous unwanted variation. The default is set to \code{3}.}

\item{nb.ncg}{Numeric. Number of genes to select as negative control genes when \code{ncg.selection.method} is \code{auto}.
This value represents a fraction of total genes. The default is set to \code{0.1}.}

\item{hvg.method}{Character. Method to select highly variable genes. Options: \code{var}, \code{mad}. \code{var} models gene variance
based on a mean-variance trend. \code{mad} uses the Median Absolute Deviation. The default is set to \code{var}.}

\item{ncg.selection.method}{Character. Method for selecting negative control genes. Options: \code{prod}, \code{sum}, \code{average},
\code{non.overlap}, \code{quantile}, \code{auto}. The default is set to \code{non.overlap}.}

\item{top.rank.bio.genes}{Numeric. Fraction of top-ranked genes highly affected by biological variation. The default
is set to \code{0.5}.}

\item{top.rank.uv.genes}{Numeric. Fraction of top-ranked genes highly affected by unwanted variation. The default is
set to \code{0.5}.}

\item{bio.percentile}{Numeric. Percentile threshold for selecting biologically variable genes (used with \code{quantile}
method). The default is set to \code{0.8}.}

\item{uv.percentile}{Numeric. Percentile threshold for selecting unwanted-variable genes (used with \code{quantile}
method). The default is set to \code{0.8}.}

\item{grid.group}{Character. Scope of grid search: \code{bio}, \code{uv}, or \code{both}.}

\item{grid.direction}{Character. Direction of grid search: \code{increase} or \code{decrease}.}

\item{grid.nb}{Numeric. Number of genes to test in grid search when \code{ncg.selection.method} is \code{auto}. The default is
set to \code{20}.}

\item{normalization}{Character. Method for library size normalization before analysis. Options: \code{CPM}, \code{TMM}, \code{VST},
\code{full}, \code{median}, \code{upper}. The default is set to \code{CPM}.}

\item{regress.out.variables}{Character. Variable(s) to regress out before analysis.}

\item{min.sample.for.mad}{Numeric. Minimum number of samples per group for MAD analysis. The default is set to \code{3}.}

\item{min.sample.for.var}{Numeric. Minimum number of samples per group for variance analysis. The default is set to \code{15}.}

\item{min.sample.for.aov}{Numeric. Minimum number of samples for ANOVA. The default is set to \code{3}.}

\item{min.sample.for.correlation}{Numeric. Minimum number of samples for correlation analysis. The default is set to \code{10}.}

\item{corr.method}{Character. Correlation method: \code{pearson} or \code{spearman}. The default is set to \code{spearman}.}

\item{a}{Numeric. Significance level (alpha) for correlation confidence intervals. The default is set to \code{0.05}.}

\item{rho}{Numeric. Hypothesized correlation value for testing. The default is set to \code{0}.}

\item{anova.method}{Character. ANOVA method: \code{aov} or \code{welch}. The default is set to \code{aov}.}

\item{filter.ncgs}{Logical. Whether to filter selected NCGs based on public human housekeeping gene sets. The default
is set to \code{FALSE}.}

\item{common.hk}{Character. Specifies group of housekeeping genes to use: \code{cancer} or \code{non.cancer}. The default is set
to \code{cancer}.}

\item{hk.group}{Character. Column name in the gene annotation containing non-cancer housekeeping genes. Options include:
\code{bulk.rnaseq.hk.genes.v1}, \code{bulk.rnaseq.hk.genes.v2}, \code{micorarray.hk.genes}, \code{nanostring.pan.cancer.hk.genes},
\code{singscore.pan.cancer.hk.genes}. The default is set to \code{micorarray.hk.genes}.}

\item{assess.ncg}{Logical. Whether to assess the performance of selected NCGs using PCA and R² or vector correlations.
The default is set to \code{TRUE}.}

\item{variables.to.assess.ncg}{Character. Variables to assess with the NCGs. The default is set to \code{NULL} (all in \code{uv.variables}).}

\item{nb.pcs}{Numeric. Number of principal components to use for NCG assessment. The default is set to \code{10}.}

\item{center}{Logical. Whether to center data before PCA. The default is set to \code{TRUE}.}

\item{scale}{Logical. Whether to scale data before PCA. The default is set to \code{TRUE}.}

\item{apply.log}{Logical. Whether to log-transform the data before analysis. The default is set to \code{TRUE}.}

\item{pseudo.count}{Numeric. Pseudo count to add before log transformation. The default is set to \code{1}.}

\item{assess.se.obj}{Logical. Whether to validate the \code{SummarizedExperiment} object with \code{checkSeObj()}. The The default
is set to \code{TRUE}.}

\item{remove.na}{Character. Whether to remove missing values from \code{assays}, \code{sample.annotation}, \code{both}, or \code{none}.
The The default is set to \code{both}.}

\item{output.name}{Character. Output name for metadata entry. If \code{NULL}, a name is auto-generated.}

\item{ncg.group}{Character. Group label for selected NCGs. If \code{NULL}, the label \code{ncg|unsupervised} is used.}

\item{plot.output}{Logical. Whether to plot NCG assessment during execution. The default is set to \code{TRUE}.}

\item{output.plot}{Character. Plot type to display: \code{assessment}, \code{heatmap}, or \code{both}. The default is set to \code{assessment}.}

\item{use.imf}{Logical. Whether to use intermediate file. The default is set to \code{FALSE}.}

\item{save.imf}{Logical. Whether to save the intermediate file for reuse. The default is set to \code{FALSE}.}

\item{imf.name}{Character. Name for intermediate file. If \code{NULL}, generated as \code{paste0(assay.name, '|un.supervised|', ncg.selection.method)}.}

\item{save.se.obj}{Logical. Whether to save results in \code{metadata} of the \code{SummarizedExperiment} object. The default is set to \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE}, function will print progress messages.}
}
\value{
A \code{SummarizedExperiment} object containing the selected negative control genes and optional assessment plots,
or a list of the results.
}
\description{
Identifies a set of genes to be used as negative control genes when no biological variation is known. This function
applies gene-level correlation, ANOVA, mean-variance relationship, and MAD analyses across and between sample groups
to select negative control genes for different purposes, including \code{RUV-III} normalization.
}
\details{
The function initially uses gene-level correlation and ANOVA to identify genes significantly influenced
by continuous and categorical sources of variation, respectively. Then, it performs a mean-variance relationship
or Median Absolute Deviation (MAD) analysis on each gene within homogeneous sample groups, considering unwanted variables,
to detect genes that are highly variable due to biological factors. Finally, various methods are applied to consolidate
the statistical findings and determine an appropriate set of genes as negative control genes for different purposes,
including \code{RUV-III} normalization.
}
\author{
Ramyar Molania
}
