% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNcgsUnSupervisedByAnovaCorr.R
\name{findNcgsUnSupervisedByAnovaCorr}
\alias{findNcgsUnSupervisedByAnovaCorr}
\title{Finds NCGs using an unsupervised approach.}
\usage{
findNcgsUnSupervisedByAnovaCorr(
  se.obj,
  assay.name,
  uv.variables,
  ncg.selection.method = "quantile",
  use.rank = FALSE,
  samples.to.use = "all",
  nb.ncg = 0.1,
  hvg.method = "mad",
  top.rank.bio.genes = 0.8,
  top.rank.uv.genes = 0.2,
  bio.percentile = 0.2,
  uv.percentile = 0.2,
  grid.group = "uv",
  grid.direction = "decrease",
  grid.nb = 40,
  clustering.method = "kmeans",
  nb.clusters = 3,
  min.sample.for.mad = 3,
  min.sample.for.var = 15,
  min.sample.for.aov = 3,
  min.sample.for.correlation = 10,
  corr.method = "spearman",
  a = 0.05,
  rho = 0,
  anova.method = "aov",
  create.ncg.rank.plot = FALSE,
  plot.ncg.rank = FALSE,
  filter.ncgs = FALSE,
  common.hk = "cancer",
  nb.stable.genes = 2000,
  hk.group = "micorarray.hk.genes",
  assess.ncg = TRUE,
  variables.to.assess.ncg = NULL,
  nb.pcs = 5,
  center = TRUE,
  scale = FALSE,
  plot.ncg.assessment = TRUE,
  regress.out.variables = NULL,
  regress.out.rle.med = FALSE,
  normalization = "CPM",
  apply.log = TRUE,
  pseudo.count = 1,
  check.se.obj = TRUE,
  remove.na = "both",
  ncg.group.name = NULL,
  ncg.set.name = NULL,
  save.imf = FALSE,
  use.imf = FALSE,
  imf.name = NULL,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{Character. A character that indicates the name of an data (assay) in the SummarizedExperiment object.
The selected assay should be the one that will be used for the RUV-III-PRPS normalization.}

\item{uv.variables}{Character. A character or a vector of characters indicating the name of the columns in the
SummarizedExperiment object that contain unwanted variables. These variables can be both categorical or continouse or
a combination of these. If all unwanted variation is unknown, use the  \code{identifyUnknownUV()} function to estimate them
first, then run \code{findNcgsUnSupervised()}.}

\item{ncg.selection.method}{Character. A character that indicates how to summarize different statistics and select a
set of genes as negative control genes. The options are: \code{prod}, \code{average}, \code{sum}, \code{non.overlap}, \code{auto}, and \code{quantile}.
The default is set to \code{non.overlap}. For more information, refer to the details of the function.}

\item{use.rank}{Logical. If \code{TRUE}, the function use the rank based approach to select NCG.}

\item{samples.to.use}{Logical. A logical vector indicating wghich samples to be used for analysis. The defualt is set
to \code{all}, this means all samples be used.}

\item{nb.ncg}{Numeric. A numeric value that specifies the number of genes to be chosen as negative control genes (NCG)
when the \code{ncg.selection.method} parameter is set to \code{auto}. This value, \code{nb.ncg}, corresponds to a fraction of the total
genes in the SummarizedExperiment object. The default is set to 0.1.}

\item{hvg.method}{Character. A character indicating the method to select highly variable genes. The options are: \code{var}
and \code{mad}. \code{var} models gene variance based on a mean-variance trend. \code{mad} uses the median absolute deviation. The
default is set to \code{var}.}

\item{top.rank.bio.genes}{Numeric. A numeric value that indicates the percentage of top-ranked genes that are highly
affected by biological variation. This is required to be specified when the \code{ncg.selection.method} is either \code{auto}
or \code{non.overlap}. The default is set to 0.2.}

\item{top.rank.uv.genes}{Numeric. A numeric value that indicates the percentage of top-ranked genes that are highly
affected by unwanted variables. This is required to be specified when the \code{ncg.selection.method} is either \code{auto} or
\code{non.overlap}. The default is set to 0.2.}

\item{bio.percentile}{Numeric. A numeric value that specifies the percentile cut-off to select genes that are highly
affected by biological variation. This is required to be specified when the \code{ncg.selection.method} is set to \code{quantile}.
The default is set to 0.8.}

\item{uv.percentile}{Numeric. A numeric value that specifies the percentile cut-off to select genes that are highly
affected by unwanted variation. This is required to be specified when the \code{ncg.selection.method} is set to \code{quantile}.
The default is set to 0.8.}

\item{grid.group}{Character. A character that indicates whether the grid search should be performed on biological,
unwanted, or both factors when the \code{ncg.selection.method} is set to \code{auto}. The options are \code{bio}, \code{uv}, or \code{both}.
The default is set to \code{uv}. For more details, refer to the function documentation.}

\item{grid.direction}{Character. A character that indicates whether the grid search should be performed in decreasing
or increasing order when the \code{ncg.selection.method} is set to \code{auto}. The options are: \code{increase} and \code{decrease}. The
default is set to \code{decrease}.}

\item{grid.nb}{Numeric. A numeric value that indicates the number of genes for grid search when the \code{ncg.selection.method}
is set to \code{auto}. In the \code{auto} approach, the grid search starts with the initial \code{top.rank.uv.genes} and
\code{top.rank.bio.genes} values and adds or drops the \code{grid.nb} in each loop to find \code{nb.ncg} of genes as negative control
genes. The default is set to 20.}

\item{clustering.method}{Character. The clustering method that should be used to group any continuous unwanted variable.
The options are: \code{kmeans}, \code{cut}, and \code{quantile}. The default is set to \code{kmeans}.}

\item{nb.clusters}{Numeric. A numeric value that indicates the number of clusters for grouping continuous unwanted
variation. The default is set to 3.}

\item{min.sample.for.mad}{Numeric. A numeric value that indicates minimum number of samples required per group for MAD
analysis. The default is set to 3.}

\item{min.sample.for.var}{Numeric. A numeric value that indicates minimum number of samples required per group for
gene-variance relationship analysis. The default is set to 15.}

\item{min.sample.for.aov}{Numeric. A numeric value that indicates minimum number of samples required per group  for
ANOVA. The default is set to 3.}

\item{min.sample.for.correlation}{Numeric. A numeric value that indicates minimum number of samples required for
correlation analysis. The default is set to 10}

\item{corr.method}{Character. A character that indicates which correlation methods should be used for the correlation
analyses. The options are \code{pearson} or \code{spearman}. The default is set to \code{spearman}.}

\item{a}{Numeric. The significance level used for the confidence intervals in the correlation; by default, it is set
to 0.05. Refer to the function \code{correls} from the \strong{Rfast} R package for more details.}

\item{rho}{Numeric. The value of the hypothesized correlation to be used in the hypothesis testing. The default is
set to 0. Refer to the function \code{correls} from the \strong{Rfast} R package for more details.}

\item{anova.method}{Character. A character that  indicates which ANOVA method to use. The options are \code{aov} or \code{welch}.
The default is se to \code{aov}. Refer to the function \code{row_oneway_equalvar()} or \code{row_oneway_welch()} from the
\strong{matrixTests} R package for more details.}

\item{create.ncg.rank.plot}{Logical. Indicates whether to generate a heatmap that shows the rank of the all genes
with respect to their biological and unwanted variation effects. The default is set to \code{FALSE}.}

\item{plot.ncg.rank}{Logical. Indicates whether to plot a heatmap that shows the rank of the all genes
with respect to their biological and unwanted variation effects, while function is running. The default is set to \code{FALSE}.}

\item{filter.ncgs}{Logical. Whether to filter selected NCGs based on public human housekeeping gene sets. The default
is set to \code{FALSE}.}

\item{common.hk}{Character. Specifies group of housekeeping genes to use: \code{cancer} or \code{non.cancer}. The default is set
to \code{cancer}.}

\item{nb.stable.genes}{Numeric. A numeric value that specifies the number of top stable genes to be obtained from the
\code{getStableGenes()} function in the \strong{singescore} R package. The default is set to 2000.}

\item{hk.group}{Character. Column name in the gene annotation containing non-cancer housekeeping genes. Options include:
\code{bulk.rnaseq.hk.genes.v1}, \code{bulk.rnaseq.hk.genes.v2}, \code{micorarray.hk.genes}, \code{nanostring.pan.cancer.hk.genes},
\code{singscore.pan.cancer.hk.genes}. The default is set to \code{micorarray.hk.genes}.}

\item{assess.ncg}{Logical. Indicates whether to assess the performance of selected genes as negative control or not.
This analysis involves principal component analysis on the selected genes, followed by exploration of the R^2 or vector
correlation between the first \code{nb.pcs} principal components and the biological and unwanted variables. The default is
set to \code{TRUE}.}

\item{variables.to.assess.ncg}{Character. A character string or vector of strings indicating the column names in sample
annotation of of the  SummarizedExperiment object that contain variables whose association with the selected genes as
NCG needs to be evaluated. The default is set to \code{NULL}. This means all the variables specified in the \code{bio.variables}
and \code{uv.variables} will be assessed.}

\item{nb.pcs}{Numeric. A numeric value that indicates the number of the first principal components of selected negative
control genes to be used to assess their performance. The default is set to 10.}

\item{center}{Logical. Indicates whether to scale the data before applying SVD. If \code{TRUE}, centering is done by subtracting
the column means of the assay from their corresponding columns. The default is set to \code{TRUE}.}

\item{scale}{Logical. Indicates whether to scale the data before applying SVD. If \code{TRUE}, scaling is done by dividing the
(centered) columns of the assays by their standard deviations if centering is \code{TRUE}, and by the root mean square otherwise.
The default is set to \code{FALSE}.}

\item{plot.ncg.assessment}{Logical. Indicates whether to plot the output of the NCG assessment while function is running
. The default is set to \code{TRUE}.}

\item{regress.out.variables}{Character. A character or a vector of character indicating the names of the columns in
the SummarizedExperiment object that contain variables to be regressed out from the data before identifying biologically
variable genes. The default is set to \code{NULL}.}

\item{regress.out.rle.med}{Logical. If \code{TRUE}, the RLE medians will be regressed out from the data in data-preprocessing
steps}

\item{normalization}{Character. A character that indicates which normalization method should be use to mitigate the
variation in library size before finding genes that are highly affected by biological variation. The options are :
\code{CPM}, \code{TMM}, \code{VST}, \code{upper}, \code{full} and \code{medium}. The default is set to  \code{CPM}. Refer to the \code{applyOtherNormalization()}
function for more details.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data before any statistical analyses.
The default is set to \code{TRUE}.}

\item{pseudo.count}{Numeric. A numeric value to be added as a pseudo count to all measurements before applying log
transformation. The default is set to 1.}

\item{check.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object before any analysis. If \code{TRUE},
the function \code{checkSeObj()} will be used. The default is set to \code{TRUE}.}

\item{remove.na}{Character set. Indicates whether to remove NA or missing values from the SummarizedExperiment object
The options are: \code{assays}, the \code{sample.annotation}, \code{both}, or \code{none}. If \code{assays} is selected, genes containing NA or
missing values will be excluded. If \code{sample.annotation} is selected, the samples containing NA or missing values for
any \code{bio.variables} or \code{uv.variables} will be excluded. The default is set to \code{none}.}

\item{ncg.group.name}{Character. A character to be used as name of the group of NCG. The default is set to \code{NULL}, then
the function create a names as following: \code{paste0('ncg|unsupervised')}. We refer to the details of the function for
more details.}

\item{ncg.set.name}{Character. A character to be used as name of the NCG set based on current variables and parameters
The default is set to \code{NULL}, then the function create a names as following:
\code{paste0(sum(ncg.selected),'|',paste0(bio.variables, collapse = '&'),'|',paste0(uv.variables, collapse = '&'),'|AnoCorrAs:', ncg.selection.method,'|',assay.name)}.We refer to the details of the function for more details.}

\item{save.imf}{Logical. Indicates whether to save the intermediate file. If \code{TRUE}, the function saves the results
of the statistical analyses in the metadata of the SummarizedExperiment object. If users want to change the parameters
including \code{nb.ncg}, \code{ncg.selection.method}, \code{top.rank.bio.genes}, and \code{top.rank.uv.genes}, the analyses will not be
re-calculated. The default is set to \code{FALSE}.}

\item{use.imf}{Logical. Indicates whether to use the intermediate file. The default is set to \code{FALSE}.}

\item{imf.name}{Character string. A name to save the intermediate file. If \code{NULL}, the function generates a name.}

\item{save.se.obj}{Logical. Indicates whether to save the result of the function in the metadata of the
SummarizedExperiment object or output the result. The default is \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE}, shows messages of different steps of the function.}
}
\value{
A \code{SummarizedExperiment} object containing the selected negative control genes and optional assessment plots,
or a list of the results.
}
\description{
This function identifies a set of genes to be used as negative control genes (NCGs) when no biological variation is
known. This function applies gene-level correlation, ANOVA, mean-variance relationship, and MAD analyses across and
between sample groups to select NCGs for different purposes, including RUV-III normalization.
}
\details{
The function initially uses gene-level correlation and ANOVA to identify genes significantly influenced by continuous
and categorical sources of variation, respectively. Then, it performs a mean-variance relationship or median absolute
deviation (MAD) analysis on each gene within homogeneous sample groups, considering unwanted variables, to detect genes
that are highly variable due to biological factors. Finally, various methods are applied to consolidate the statistical
findings and determine an appropriate set of genes as NCGs for different purposes,
including RUV-III normalization.
}
\author{
Ramyar Molania
}
