% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNcgSupervisedByLinearMixedModel.R
\name{findNcgSupervisedByLinearMixedModel}
\alias{findNcgSupervisedByLinearMixedModel}
\title{Find NCGs using linear mixed model}
\usage{
findNcgSupervisedByLinearMixedModel(
  se.obj,
  assay.name,
  bio.variables,
  uv.variables,
  form,
  use.rank = FALSE,
  ncg.selection.method = "quantile",
  adjust.data = FALSE,
  adjustment.method = "lm",
  adjustment.variables = "uv",
  samples.to.use = "all",
  nb.ncg = 0.1,
  top.rank.bio.genes = 0.5,
  top.rank.uv.genes = 0.5,
  bio.percentile = 0.2,
  uv.percentile = 0.8,
  grid.group = "uv",
  grid.direction = "increase",
  grid.nb = 20,
  apply.log = TRUE,
  pseudo.count = 1,
  assess.ncg = TRUE,
  variables.to.assess.ncg = NULL,
  nb.pcs = 10,
  center = TRUE,
  scale = FALSE,
  svd.bsparam = bsparam(),
  plot.ncg.assessment = TRUE,
  create.ncg.rank.plot = FALSE,
  nb.cores = NULL,
  check.se.obj = TRUE,
  remove.na = "none",
  ncg.group.name = NULL,
  ncg.set.name = NULL,
  save.imf = FALSE,
  imf.name = NULL,
  use.imf = FALSE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{Character. A character string indicating the name of the data (assay) in the \code{SummarizedExperiment}
object. This data should be the one that will be used as input data for the RUV-III normalization.}

\item{bio.variables}{Character. A character string or vector of strings indicating the column name(s) of the biological
variable(s) in the SummarizedExperiment object. These variable can be categorical or continuous or a combination. This
argument cannot be \code{NULL}.}

\item{uv.variables}{Character. A character string or vector of strings indicating the column name(s) of the unwanted
variable(s) in the SummarizedExperiment object. These variable can be categorical or continuous or a combination.This
argument cannot be \code{NULL}.}

\item{form}{TTTT}

\item{use.rank}{TTTT}

\item{ncg.selection.method}{Character. A character that indicates how to summarize different statistics and select a
set of genes as negative control genes. The options are: \code{prod}, \code{average}, \code{sum}, \code{non.overlap}, \code{auto}, and \code{quantile}.
The default is set to \code{non.overlap}. For more information, refer to the details of the function.}

\item{adjust.data}{TTTT}

\item{adjustment.method}{TTTT}

\item{adjustment.variables}{TTTT}

\item{samples.to.use}{TTTT}

\item{nb.ncg}{Numeric. A numeric value that specifies the number of genes to be chosen as negative control genes (NCG)
when the \code{ncg.selection.method} parameter is set to \code{auto}. This value, \code{nb.ncg}, corresponds to a fraction of the total
genes in the SummarizedExperiment object. The default is set to 0.1.}

\item{top.rank.bio.genes}{Numeric. A numeric value that indicates the percentage of top-ranked genes that are highly
affected by biological variation. This is required to be specified when the \code{ncg.selection.method} is either \code{auto}
or \code{non.overlap}. The default is set to 0.2.}

\item{top.rank.uv.genes}{Numeric. A numeric value that indicates the percentage of top-ranked genes that are highly
affected by unwanted variables. This is required to be specified when the \code{ncg.selection.method} is either \code{auto} or
\code{non.overlap}. The default is set to 0.2.}

\item{bio.percentile}{Numeric. A numeric value that specifies the percentile cut-off to select genes that are highly
affected by biological variation. This is required to be specified when the \code{ncg.selection.method} is set to \code{quantile}.
The default is set to 0.8.}

\item{uv.percentile}{Numeric. A numeric value that specifies the percentile cut-off to select genes that are highly
affected by unwanted variation. This is required to be specified when the \code{ncg.selection.method} is set to \code{quantile}.
The default is set to 0.8.}

\item{grid.group}{Character. A character that indicates whether the grid search should be performed on biological,
unwanted, or both factors when the \code{ncg.selection.method} is set to \code{auto}. The options are \code{bio}, \code{uv}, or \code{both}.
The default is set to \code{uv}. For more details, refer to the function documentation.}

\item{grid.direction}{Character. A character that indicates whether the grid search should be performed in decreasing
or increasing order when the \code{ncg.selection.method} is set to \code{auto}. The options are: \code{increase} and \code{decrease}. The
default is set to \code{decrease}.}

\item{grid.nb}{Numeric. A numeric value that indicates the number of genes for grid search when the \code{ncg.selection.method}
is set to \code{auto}. In the \code{auto} approach, the grid search starts with the initial \code{top.rank.uv.genes} and
\code{top.rank.bio.genes} values and adds or drops the \code{grid.nb} in each loop to find \code{nb.ncg} of genes as negative control
genes. The default is set to 20.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data before performing any statistical
analysis. The default is set to \code{TRUE}.}

\item{pseudo.count}{Numeric. A numeric value to be added as a pseudo count to all measurements before log transformation
.The default is se to 1.}

\item{assess.ncg}{Logical. Indicates whether to assess the performance of selected genes as negative control or not.
This analysis involves principal component analysis on the selected genes, followed by exploration of the R^2 or vector
correlation between the first \code{nb.pcs} principal components and the biological and unwanted variables. The default is
set to \code{TRUE}.}

\item{variables.to.assess.ncg}{Character. A character string or vector of strings indicating the column names in sample
annotation of of the  SummarizedExperiment object that contain variables whose association with the selected genes as
NCG needs to be evaluated. The default is set to \code{NULL}. This means all the variables specified in the \code{bio.variables}
and \code{uv.variables} will be assessed.}

\item{nb.pcs}{Numeric. A numeric value that indicates the number of the first principal components of selected negative
control genes to be used to assess their performance. The default is set to 10.}

\item{center}{Logical. Indicates whether to scale the data before applying SVD. If \code{TRUE}, centering is done by subtracting
the column means of the assay from their corresponding columns. The default is set to \code{TRUE}.}

\item{scale}{Logical. Indicates whether to scale the data before applying SVD. If \code{TRUE}, scaling is done by dividing
the(centered) columns of the assays by their standard deviations if centering is \code{TRUE}, and by the root mean square
otherwise. The default is set to \code{FALSE}.}

\item{svd.bsparam}{TTTT}

\item{plot.ncg.assessment}{Logical. Indicates whether to plot the output of the NCG assessment while function is running
. The default is set to \code{TRUE}.}

\item{create.ncg.rank.plot}{Logical. Indicates whether to generate a heatmap that shows the rank of the all genes
with respect to their biological and unwanted variation effects. The default is set to \code{FALSE}.}

\item{nb.cores}{Numeric.}

\item{check.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object before any analysis. If \code{TRUE},
the function \code{checkSeObj()} will be used. The default is set to \code{TRUE}.}

\item{remove.na}{Character set. Indicates whether to remove NA or missing values from the SummarizedExperiment object
The options are: \code{assays}, the \code{sample.annotation}, \code{both}, or \code{none}. If \code{assays} is selected, genes containing NA or
missing values will be excluded. If \code{sample.annotation} is selected, the samples containing NA or missing values for
any \code{bio.variables} or \code{uv.variables} will be excluded. The default is set to \code{none}.}

\item{ncg.group.name}{Character. A character to be used as name of the group of NCG. The default is set to \code{NULL}, then
the function create a names as following: \code{paste0('ncg|unsupervised')}. We refer to the details of the function for
more details.}

\item{ncg.set.name}{Character. A character to be used as name of the NCG set based on current variables and parameters
The default is set to \code{NULL}, then the function create a names as following:
\code{paste0(sum(ncg.selected),'|',paste0(bio.variables, collapse = '&'),'|',paste0(uv.variables, collapse = '&'),'|AnoCorrAs:', ncg.selection.method,'|',assay.name)}.We refer to the details of the function for more details.}

\item{save.imf}{Logical. Indicates whether to save the intermediate file. If \code{TRUE}, the function saves the results
of the statistical analyses in the metadata of the SummarizedExperiment object. If users want to change the parameters
including \code{nb.ncg}, \code{ncg.selection.method}, \code{top.rank.bio.genes}, and \code{top.rank.uv.genes}, the analyses will not be
re-calculated. The default is set to \code{FALSE}.}

\item{imf.name}{Character string. A name to save the intermediate file. If \code{NULL}, the function generates a name.}

\item{use.imf}{Logical. Indicates whether to use the intermediate file. The default is set to \code{FALSE}.}

\item{save.se.obj}{Logical. Indicates whether to save the result of the function in the metadata of the SummarizedExperiment
object or output the result. The default is \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE}, shows messages of different steps of the function.}
}
\description{
This function utilizes two-way ANOVA to identify a set of suitable genes as negative control genes (NCG) for RUV-III
normalization. Both biological and unwanted variation sources is necessary and should be specified.
}
\details{
The function begins by creating all possible sample groups based on biological and unwanted variation separately.
Subsequently, these groups are used as factors in two-way ANOVA to identify genes highly influenced by biological and
unwanted variation. Finally, the function selects genes with the possible highest F-statistics for unwanted variation and
lowest F-statistics for biological variation. Various approaches are employed for the final gene selection; please refer
to the details for more information.
The function uses 5 ways to summarize two gene-level F-statistics obtained for the biological and unwanted variation.
The function uses either the values or the ranks of F-statistics for NCGs selection. The function ranks the
negative of F-statistics values for unwanted variation. The lower the ranks, the greater the impact of unwanted
variation on genes. The function ranks the F-statistics for biological variation. The higher the ranks, the greater
the impact of biological variation on genes. The options are \code{prod}, \code{sum}, \code{average}, \code{auto}, \code{non.overlap} and
\code{quantile}.

If \code{prod}, \code{sum} and \code{average} is set:
\itemize{
\item The product, sum or average of ranks of F-statistics is calculated. Then, the function selects \code{nb.ncg} numbers of
genes as negative control genes that have the lowest ranks.
}

If \code{non.overlap} is selected:
\enumerate{
\item The function selects the top \code{top.rank.bio.genes} genes that have the highest ranks of F-statistics
for biological variation.
\item The function selects the top \code{top.rank.uv.genes} genes that have the lowest ranks of F-statistics for
unwanted variation.
\item The function excludes all genes obtained in 2 from the ones obtained 1. This will be a set of genes as
negative control genes.
}

If \code{auto} is selected:
\enumerate{
\item The function selects the top \code{top.rank.bio.genes} genes that have the highest ranks of F-statistics for
biological variation.
\item The function selects the top \code{top.rank.uv.genes} genes that have the lowest ranks of F-statistics for
unwanted variation.
\item The function excludes all genes obtained in 2 from the ones obtained 1.
\item If the number of selected genes is larger or smaller than the specified \code{nb.ncg}, the function applies an
auto search to find approximate \code{nb.ncg} of genes as negative control genes as follow. The auto search will either
decrease or increase the values of either \code{top.rank.bio.genes} or \code{top.rank.uv.genes} or both till to find
approximate \code{nb.ncg} of genes as negative control genes.
}

If \code{quantile} is selected:
\enumerate{
\item The function selects the \code{bio.percentile} percentile of F-statistics for biological variation. Then, selects
all the genes that have F-statistics larger the calculated percentile.
\item The function selects the \code{uv.percentile} percentile of F-statistics for unwanted variation. Then, selects
all the genes that have F-statistics larger the calculated percentile.
\item The function excludes all genes obtained in 2 from the ones obtained 1.
}

Assess the performance of NCGS:
\itemize{
\item The function can assess the initial performance of selected NCGs. This analysis involves principal component analysis
on only the selected NCG and then explore the R^2 or vector correlation between the \code{nb.pcs} first principal components
and with the specified variables. Ideal NCGS, should show high and low R^2 or vector correlation for unwanted and
biological variation respectively.
}
}
\references{
\itemize{
\item Gandolfo L. C. & Speed, T. P., RLE plots: visualizing unwanted variation in high dimensional data. PLoS ONE, 2018.
\item Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
}
\author{
Ramyar Molania
}
