% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssessmentMetrics.R
\name{getAssessmentMetrics}
\alias{getAssessmentMetrics}
\title{Creates all possible assessment plots and metrics for the variables.}
\usage{
getAssessmentMetrics(se.obj, variables, plot.output = TRUE, save.se.obj = TRUE)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{variables}{Character string or vector. Specifies the column name(s) of variables in the sample annotation
of the SummarizedExperiment object. The \code{variables} can be either categorical or continuous.}

\item{plot.output}{Logical. Whether to display the plot of all possible assessment plots. Default is \code{TRUE}.}

\item{save.se.obj}{Logical. Whether to save the results into the SummarizedExperiment object. Default is \code{TRUE}.}
}
\value{
A list of all possible assessment metrics for the variables.
}
\description{
This function provides the names of all possible assessment plots and metrics for the given variable(s). The list
can be used in the \code{assessVariation} and \code{assessNormalization} functions.
}
\author{
Ramyar Molania
}
