% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetricFromSeObj.R
\name{getMetricFromSeObj}
\alias{getMetricFromSeObj}
\title{Get a metric or plot from the \code{metadata} of a SummarizedExperiment object.}
\usage{
getMetricFromSeObj(
  se.obj,
  slot = "Metrics",
  assay.names,
  assessment.type = "gene.level",
  assessment,
  method = NULL,
  variables,
  file.name,
  sub.file.name = NULL,
  required.function,
  message.to.print,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{slot}{Character string. Indicates the name of the slot in the \code{metadata} of the SummarizedExperiment object.}

\item{assay.names}{Character string or vector. Specifies the name(s) of the assay(s) in the SummarizedExperiment
object. Default is \code{all}.}

\item{assessment.type}{Character string. Specifies the type of assessment. Options are \code{gene.level} or \code{global.level}.}

\item{assessment}{Character string. Name of the metric to retrieve.}

\item{method}{Character string. Method used to calculate the metric.}

\item{variables}{Character string or vector. Specifies the variable(s) used to calculate the metric.}

\item{file.name}{Character string. Name of the file to which the metric results were saved.}

\item{sub.file.name}{Character string. Name of the sub-file (if applicable) to which the results were saved.}

\item{required.function}{Character string. Name of the function that must have been run prior to retrieving this metric.}

\item{message.to.print}{Character string. A message to print when retrieving the metric.}

\item{verbose}{Logical. If \code{TRUE}, prints messages describing the function's progress.}
}
\description{
Get a metric or plot from the \code{metadata} of a SummarizedExperiment object.
}
\author{
Ramyar Molania
}
