% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtainMetric.R
\name{obtainMetric}
\alias{obtainMetric}
\title{Computes ANOVA between individual gene expression and a categorical variable}
\usage{
obtainMetric(
  se.obj,
  slot = "Metrics",
  metric.group = "gene.level",
  assay.name,
  metric.name,
  test.name,
  variable,
  show.path = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{slot}{TTTT}

\item{metric.group}{TTTT}

\item{assay.name}{TTTT}

\item{metric.name}{TTTT}

\item{test.name}{TTTT}

\item{variable}{TTTT}

\item{show.path}{TTTT}
}
\value{
Either a SummarizedExperiment object containing the log2 F-statistics and p-values of ANOVA for the continuous
variable or a list of these results.
}
\description{
This function calculates the ANOVA between individual gene expression of the assay(s) in a SummarizedExperiment object
and a categorical variable as factor.
}
\author{
Ramyar Molania
}
