% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderSeObj.R
\name{orderSeObj}
\alias{orderSeObj}
\title{Order SummarizedExperiment object based on variables}
\usage{
orderSeObj(se.obj, factors.to.order)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{factors.to.order}{TTT}
}
\value{
The function returns a log2 transformed of all specified data sets as a list object.
}
\description{
Order SummarizedExperiment object based on variables
}
\author{
Ramyar Molania
}
