% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam50.robust.R
\docType{data}
\name{pam50.robust}
\alias{pam50.robust}
\title{PAM50 Robust List}
\format{
A list with X elements:
\describe{
\item{method.cor}{Description of element1}
\item{method.centroids}{Description of element2}
\item{rescale.q}{Description of element3}
\item{mins}{Description of element3}
\item{centroids}{Description of element3}
\item{centroids.map}{Description of element3}
...
}
}
\usage{
pam50.robust
}
\description{
A list containing robust PAM50 gene signature objects used for
molecular subtyping in prostate cancer analysis.
}
\examples{
data(pam50.robust)
names(pam50.robust)
}
\keyword{datasets}
