% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotARI.R
\name{plotARI}
\alias{plotARI}
\title{Generate barplots or scatter plots the Adjusted Rand Index (ARI).}
\usage{
plotARI(
  se.obj,
  assay.names = "all",
  variables,
  ari.method = "hclust.complete.euclidian",
  plot.type = "single.plot",
  plot.output = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Character or character vector. Specifies the name(s) of the assay(s) in the \code{SummarizedExperiment}
object to be used for generating barplots or scatter plots of the computed Adjusted Rand Index (ARI). By default,
all assays in the \code{SummarizedExperiment} object will be selected.}

\item{variables}{Character or character vector of length one or two. Indicates one or two column names in the
\code{SummarizedExperiment} object that contain categorical variables, such as sample subtypes or batch labels. If two
variables are provided, the function plots the ARIs against each other for all specified assays.}

\item{ari.method}{Character. Specifies the ARI computation method to use for plotting. This must match one of the
methods supported by the \code{computeARI} function. The default is \code{"hclust.complete.euclidean"}, which is also the default
in \code{computeARI}. See the \code{computeARI} function documentation for more details.}

\item{plot.type}{Character. Specifies the plot type for the Adjusted Rand Index. Options are \code{"single.plot"} and
\code{"combined.plot"}. If one variable is provided in \code{variables}, \code{plot.type} must be set to \code{"single.plot"} to generate
a barplot. If two variables are provided and \code{plot.type} is set to \code{"combined.plot"}, the function generates a scatter
plot comparing the ARIs. The default is set to \code{"single.plot"}.}

\item{plot.output}{Logical. If \code{TRUE}, the individual barplots or scatter plots will be printed during function execution.
The default is to \code{TRUE}.}

\item{save.se.obj}{Logical. Indicates whether to save the plots in the metadata of the \code{SummarizedExperiment} object
or return them as a list. The default is \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE}, messages for different steps of the function will be displayed.}
}
\value{
A \code{SummarizedExperiment} object or a list containing all the plots of the computed ARIs for the categorical
variable(s).
}
\description{
This function generates barplots of the Adjusted Rand Index (ARI) for individual datasets in the \code{SummarizedExperiment}
object. If two variables are provided, the function creates scatter plots comparing the ARIs of each variable across
the assays. The \code{computeARI()} function must applied before using the \code{plotARI()} function.
}
\references{
Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
\author{
Ramyar Molania
}
