% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAssessVariation.R
\name{plotAssessVariation}
\alias{plotAssessVariation}
\title{Plot the results of the assess variation .}
\usage{
plotAssessVariation(
  se.obj,
  assay.names = "all",
  variables,
  fast.pca = TRUE,
  anova.method = "aov",
  corr.method = "spearman",
  pcorr.method = "spearman",
  sil.dist.measure = "euclidian",
  ari.clustering.method = "hclust",
  ari.hclust.method = "complete",
  ari.hclust.dist.measure = "euclidian",
  output.file.name = NULL,
  pdf.width = 12,
  pdf.height = 12,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Character or character vector. Specifies the name(s) of the assay(s) in the
SummarizedExperiment object. The default is set to \code{all}, so all the assays in the SummarizedExperiment object will
be selected.}

\item{variables}{Character or character vector. Specifies the name(s) of the column(s) in the SummarizedExperiment
object for which the specified metrics below have been calculated.}

\item{fast.pca}{Logical. Indicates whether the fast SVD approach was used in the \code{computePCA} function. The default
is set to \code{TRUE}.}

\item{anova.method}{Character. Indicates which ANOVA method was used to compute the association between
gene-level expression and a categorical variable. Options are: \code{aov} and \code{welch}. Default is \code{aov}.}

\item{corr.method}{Character. Specifies which correlation method to use to compute correlation between
gene-level expression and a continuous variable. Options are: \code{pearson}, \code{kendall}, or \code{spearman}. Default is \code{spearman}.}

\item{pcorr.method}{Character. Indicates which correlation method to use to compute gene-gene partial
correlation. Options are: \code{pearson}, \code{kendall}, or \code{spearman}. Default is \code{spearman}.}

\item{sil.dist.measure}{Character. Indicates which distance measure to apply on the PCs to calculate silhouette scores.
Options are: \code{euclidean}, \code{maximum}, \code{manhattan}, \code{canberra}, \code{binary}, or \code{minkowski}. Default is \code{euclidean}.
Refer to the function \code{computeSilhouette} for more details.}

\item{ari.clustering.method}{Character. Indicates which clustering method to apply on the PCs to
calculate ARI. Options are: \code{mclust} or \code{hclust}. Default is \code{hclust}. Refer to the \code{computeARI} function for more details.}

\item{ari.hclust.method}{Character. Specifies the agglomeration method to use when \code{clustering.method}
is set to \code{hclust} for computing ARI. Options are: \code{ward.D}, \code{ward.D2}, \code{single}, \code{complete}, \code{average} (UPGMA),
\code{mcquitty} (WPGMA), \code{median} (WPGMC), or \code{centroid} (UPGMC). Default is \code{complete}. Refer to the \code{computeARI}
function for more details.}

\item{ari.hclust.dist.measure}{Character. Specifies the distance measure to use in the \code{dist} function when
\code{clustering.method} is set to \code{hclust} for computing ARI. Options are: \code{euclidean}, \code{maximum}, \code{manhattan},
\code{canberra}, \code{binary}, or \code{minkowski}. Default is \code{euclidean}. Refer to the \code{computeARI} function for more details.}

\item{output.file.name}{Character. The path and name of the output file where the assessment plots will be saved in
PDF format.}

\item{pdf.width}{Numeric. Specifies the width (in inches) of the output PDF file. Default is typically around 8.}

\item{pdf.height}{Numeric. Specifies the height (in inches) of the output PDF file. Default is typically around 6.}

\item{verbose}{Logical. If \code{TRUE}, displays messages describing different steps of the function.}
}
\description{
Plot the results of the assess variation .
}
