% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDGE.R
\name{plotDGE}
\alias{plotDGE}
\title{Plot histograms of p-values from differential gene expression analysis}
\usage{
plotDGE(
  se.obj,
  assay.names = "all",
  variable,
  method = "limma",
  plot.ncol = 1,
  plot.nrow = 2,
  plot.output = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A \code{SummarizedExperiment} object containing expression
data and sample-level metadata.}

\item{assay.names}{Character vector specifying the name(s) of assay(s) in
the \code{SummarizedExperiment} object to be used for DGE analysis.
The default is \code{"all"}, indicating that all available assays will be
evaluated.}

\item{variable}{Character string specifying the column name in
\code{colData(se.obj)} containing the categorical variable of interest
(e.g., batch, condition, or sample type).}

\item{method}{Character string specifying the differential expression method
to be used.}

\item{plot.ncol}{Numeric. Number of columns in the plot grid when multiple
datasets are provided (default: 1).}

\item{plot.nrow}{Numeric. Number of rows in the plot grid when multiple
datasets are provided (default: 1).}

\item{plot.output}{Logical. If \code{TRUE}, p-value histograms are displayed
during execution. Default is \code{FALSE}.}

\item{save.se.obj}{Logical. If \code{TRUE}, the results are stored in the
metadata of the input \code{SummarizedExperiment} object and returned.
If \code{FALSE}, the function returns the results as a separate object.
Default is \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE}, informative messages describing
function progress are printed.}
}
\value{
A \code{SummarizedExperiment} object with appended metadata or a
list containing the computed assessment results.
}
\description{
Generates histograms of p-values obtained from differential gene expression
(DGE) analyses to assess the presence and impact of unwanted sources of
variation, such as batch effects, across multiple assays.
}
\details{
Differential gene expression (DGE) analysis is performed using the Wilcoxon
signed-rank test on log-transformed expression values (e.g., raw counts,
normalized data, or batch-corrected data). To evaluate the influence of
unwanted variation, DGE analysis is conducted across batches or other
categorical variables. In the absence of systematic bias or batch effects,
the histogram of unadjusted p-values is expected to follow a uniform
distribution. Deviations from uniformity may indicate the presence of
technical artifacts or confounding factors.
}
\author{
Ramyar Molania
}
