% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneSetScore.R
\name{plotGeneSetScore}
\alias{plotGeneSetScore}
\title{Plots the score of the gene set enrichment analysis.}
\usage{
plotGeneSetScore(
  se.obj,
  assay.names = "all",
  reference.score = NULL,
  gene.set.name,
  plot.output = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Character. A character or character vector that specifies the name(s) of the data (assays) in the
SummarizedExperiment object. The specified \code{gene.set.name} must be calculated by the \code{computeGeneSetScore()} function
before. The default is set to \code{all}, indicating all data sets in the SummarizedExperiment object will be selected.}

\item{reference.score}{Character. A character string specifying the name of the data set whose gene set score will be
used as the reference. All scores from other data sets will be plotted against this reference. The default is set to
\code{NULL}, indicating that scores from all data sets will be plotted against each other.}

\item{gene.set.name}{Character. A character string indicating the name of the gene set stored in the metadata of the
SummarizedExperiment object. If \code{NULL}, the function will select the gene set name generated using the default parameters
of the \code{computeGeneSetScore()} function. See \code{computeGeneSetScore()} for more details.}

\item{plot.output}{Logical. Whether to plot the gene set score or not. The default is set to \code{TRUE}.}

\item{save.se.obj}{Logical. Indicates whether to save the plots of scores in the metadata of the SummarizedExperiment
object or to output the result as a list. The default is set to \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE}, shows the messages of different steps of the function.}
}
\description{
This function plots the score of the gene set scoring analysis performed by the \code{computeGeneSetScore()} function.
}
\author{
Ramyar Molania
}
