% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenesPartialCorrelation.R
\name{plotGenesPartialCorrelation}
\alias{plotGenesPartialCorrelation}
\title{Plot gene-gene partial correlation.}
\usage{
plotGenesPartialCorrelation(
  se.obj,
  assay.names = "all",
  variable = NULL,
  method = "spearman",
  plot.type = "barplot",
  filter.genes = TRUE,
  corr.dif.cutoff = 0.4,
  plot.ncol = 2,
  plot.nrow = 3,
  plot.output = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Character. A character string or a vector of character strings specifying the names of the assays
in the SummarizedExperiment object for which the computed correlations are to be obtained. The default is set to "all,"
indicating that all assays of the SummarizedExperiment object will be selected.}

\item{variable}{Character. A character string indicating the column name in the SummarizedExperiment object that was used
to compute partial correlation.}

\item{method}{Character. A character string specifying the correlation method used for the "variable". The options
are 'pearson' or 'spearman'. The default is set to 'spearman'.}

\item{plot.type}{Character. A character string specifying how to plot the results of the partial pairwise gene correlation.
The options are: 'scatter plot', 'bar plot', and 'histogram'. The default is set to 'bar plot'.}

\item{filter.genes}{Logical. Indicates whether to filter genes before generating the plots. The default is set to 'TRUE'.}

\item{corr.dif.cutoff}{Numeric. A numeric value to be used as a cutoff for filtering genes. The default is set to
0.4.}

\item{plot.ncol}{Numeric. Specifies the number of columns in the plot grid. When more than one assay is selected,
the function arranges all the RLE boxplots in one grid.}

\item{plot.nrow}{Numeric. Specifies the number of rows in the plot grid. When more than three assays are selected, the function
arranges all the RLE boxplots in one grid.}

\item{plot.output}{Logical. Indicates whether to print the histogram of partial correlation coefficients. The default is
set to 'FALSE'.}

\item{save.se.obj}{Logical. Indicates whether to save the RLE results in the metadata of the SummarizedExperiment object or
output the result as a list. By default, this is set to 'TRUE'.}

\item{verbose}{Logical. If 'TRUE', displays messages for different steps of the function.}
}
\value{
The SummarizedExperiment object containing the correlation plots or a list of correlation plots
for individual assays.
}
\description{
This function computes all possible gene-gene pairwise ordinary and partial correlation of the assays in the
SummarizedExperiment object.
}
\details{
Partial correlation is used to estimate correlation between two variables while controlling for third
variables.
}
\author{
Ramyar Molania
}
