% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneVariableAnova.R
\name{plotGenesVariableAnova}
\alias{plotGenesVariableAnova}
\title{Plot F-statistics obtained from ANOVA.}
\usage{
plotGenesVariableAnova(
  se.obj,
  assay.names = "all",
  variable,
  anova.method = "aov",
  plot.ncol = 4,
  plot.nrow = 1,
  plot.output = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A \code{SummarizedExperiment} object.}

\item{assay.names}{Character or vector. A character string or a vector of strings specifying the name(s) of the
assay(s) in the \code{SummarizedExperiment} object to include. By default, all assays will be selected—provided that
ANOVA results have already been computed for them using the \code{computeGenesVariableANOVA()} function.
The function will generate plots for the ANOVA results corresponding to each selected assay.}

\item{variable}{Character. A string indicating the column name in the \code{SummarizedExperiment} object that contains
a categorical variable (e.g., sample types or batch labels). The gene-level ANOVA with this variable must
have been previously computed using the \code{computeGenesVariableANOVA()} function.}

\item{anova.method}{Logical. Indicates which ANOVA method results should be used for plotting. The options are \code{aov} or
\code{welch}. The default is set to \code{aov}.}

\item{plot.ncol}{Numeric. The number of columns in the plot grid. When more than one assay is selected, the function
arranges the plots in a grid accordingly. The default is set to 4.}

\item{plot.nrow}{Numeric. The number of rows in the plot grid. When more than three assays are selected, the function
arranges the plots in a grid accordingly. The default is set to 1.}

\item{plot.output}{Logical. If \code{TRUE}, displays the boxplot(s) of the ANOVA results. The default is set to \code{FALSE}.}

\item{save.se.obj}{Logical. If \code{TRUE}, saves the plots in the metadata of the \code{SummarizedExperiment} object. If \code{FALSE},
returns the plots as a list. The default is set to \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE}, displays messages during execution.}
}
\value{
Either a \code{SummarizedExperiment} object containing the plots in its metadata, or a list of boxplots showing
ANOVA results for the specified categorical variable across the selected assays.
}
\description{
This functions plots the results of the ANOVA analysis computed by the \code{computeGenesVariableANOVA()} function.
}
\author{
Ramyar Molania
}
