% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLisi.R
\name{plotLisi}
\alias{plotLisi}
\title{Generate barplots the local inverse Simpson's index (lisi).}
\usage{
plotLisi(
  se.obj,
  assay.names = "all",
  variable,
  fast.pca = TRUE,
  plot.output = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Character or character vector. Specifies the name(s) of the dataset(s) in the \code{SummarizedExperiment}
object to be used for generating barplots or scatter plots of the computed LISI By default, all dataset(s) in the
\code{SummarizedExperiment} object will be selected.}

\item{variable}{Character. A character specifying a column name in the sample annotation of the SummarizedExperiment
object that corresponds to the variable for which LISI has been computed.}

\item{fast.pca}{Character. A character specifying which PCA approach has been used to compute LISI The default is set
to \code{fast.pca}. Refer to the \code{computeLisi} function for more information.}

\item{plot.output}{Logical. If \code{TRUE}, the individual barplots or scatter plots will be printed during function execution.
The default is set to \code{TRUE}.}

\item{save.se.obj}{Logical. Indicates whether to save the plots in the metadata of the \code{SummarizedExperiment} object
or return them as a list. The default is set to \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE}, messages for different steps of the function will be displayed.}
}
\value{
A \code{SummarizedExperiment} object or a list containing all the plots of the computed LISI scores for the
categorical variable(s).
}
\description{
This function generates barplots of the local inverse Simpson's index (lisi) for individual dataset(s) in the
\code{SummarizedExperiment} object. Before applying the function, the \code{computeLisi} function should have been applied on the
specified dataset(s).
}
\author{
Ramyar Molania
}
