% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCVariableCorrelation.R
\name{plotPCVariableCorrelation}
\alias{plotPCVariableCorrelation}
\title{Generate line-dot plot of the PCs-variable vector correlation analysis.}
\usage{
plotPCVariableCorrelation(
  se.obj,
  assay.names = "all",
  variable,
  fast.pca = TRUE,
  nb.pcs = 10,
  plot.output = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Character. A character string or vector of character strings for the selection of the name(s) of the
assay(s) in the SummarizedExperiment object to generate a vector correlation. The default is \code{all}, which indicates all
the assays of the SummarizedExperiment object will be selected.}

\item{variable}{Character. A character string that indicates the name of the column in the sample annotation of the
SummarizedExperiment object. The variable must be a categorical variable.}

\item{fast.pca}{Logical. Indicates whether to use the fast PCA or PCA results computed by the \code{computePCA} function.
The default is set to \code{TRUE}.}

\item{nb.pcs}{Numeric. A numeric value indicating the number of first PCs to use to plot the vector correlation. The
default is set to 10.}

\item{plot.output}{Logical. If \code{TRUE}, the vector correlation line-dot plot of individual assay(s) will be printed
while the function is running.}

\item{save.se.obj}{Logical. Indicates whether to save the vector correlation plots to the meta data of the
SummarizedExperiment object or to output the result as a list. The default it is set to \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE}, displaying process messages is enabled.}
}
\value{
A SummarizedExperiment object or a list that contains all the vector correlation plots for the individual
assay(s).
}
\description{
This function generates a dot-line plot between the first cumulative PCs and the correlation coefficient (R2) obtained
from the vector correlation analysis. An ideal normalization should result in a low correlation with unwanted variation
variables and a high correlation with known biology.
}
\references{
Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
\author{
Ramyar Molania
}
