% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCVariableRegression.R
\name{plotPCVariableRegression}
\alias{plotPCVariableRegression}
\title{Generates line-dot plot of the PCs-variable regression analysis.}
\usage{
plotPCVariableRegression(
  se.obj,
  assay.names = "all",
  variable,
  fast.pca = TRUE,
  nb.pcs = 10,
  plot.output = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Character. A character string or vector of character strings specifying the name(s) of the assay(s)
in the SummarizedExperiment object to generate a vector correlation. The default is "all", which indicates that all
assays in the SummarizedExperiment object will be selected.}

\item{variable}{Character. A character string indicating the name of the column in the sample annotation
(i.e., \code{colData}) of the SummarizedExperiment object. The variable must be a continuous variable, such as library size,
tumor purity, etc.}

\item{fast.pca}{Logical. Indicates whether to use PCA calculated using a specific number of principal components
to speed up the process. The default this is set to \code{TRUE}.}

\item{nb.pcs}{Numeric. A numeric value specifying the number of leading principal components to use when plotting the
R-squared values. The default is set to 10.}

\item{plot.output}{Logical. If \code{TRUE}, the correlation statistics will be plotted. Default is TRUE.}

\item{save.se.obj}{Logical. Indicates whether to save the results in the metadata of the SummarizedExperiment object
or to return the results directly. By default, this is set to \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE}, process messages will be displayed.}
}
\value{
A SummarizedExperiment object containing the line-dot plots for the continuous variable and if requested
the associated plot.
}
\description{
This function generate a dot-line plot between the first n cumulative PCs and the regression R squared obtained
from the regression analysis. An ideal normalization should results a low R squared with unwanted variation
variables and high R squared with known biology.
}
\author{
Ramyar Molania
}
