% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRleVariable.R
\name{plotRleVariable}
\alias{plotRleVariable}
\title{Plots a variable against the medians and IQRs of relative log expression (RLE) plot.}
\usage{
plotRleVariable(
  se.obj,
  assay.names = "all",
  variable,
  rle.data.type = "both",
  ylim.rle.med.plot = NULL,
  ylim.rle.iqr.plot = NULL,
  points.size = 1,
  plot.ncol = 3,
  plot.nrow = 3,
  plot.output = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Character. A character string or a vector of character strings representing the names of the data
set(s) in the SummarizedExperiment object. The RLE data of these data set(s) must be computed before using the \code{computeRLE()}
function. The default is set to \code{all}, which indicates the RLE data of that all data set(s) in the SummarizedExperiment
object will be selected and used.}

\item{variable}{Character. A character that specifies a name of the columns in the sample annotation of the SummarizedExperiment
object. The variable can be either categorical or continuous. If a categorical variable is provided, boxplots of the
RLE medians and IQRs will be generated across the variable, separately. If a continuous variable is provided, scatter
plots of the RLE medians and IQRs will be produced against the variable, separately.}

\item{rle.data.type}{Character. Indicates which RLE data should be used for plotting. The options are \code{rle.medians},
\code{rle.iqr}, or \code{both}. If \code{rle.medians} is selected, the RLE medians will be plotted against the variable. If \code{rle.iqr}
is selected, the RLE IQRs will be plotted against the variable, and if \code{both}, both RLE medians and IQRs will be plotted
against the variable. The default is set to \code{both}.}

\item{ylim.rle.med.plot}{Numeric. Specifies the ylim of the boxplot or scatter plots when the RLE medians are used. If
\code{NULL}, the function will automatically determine an appropriate ylim for the plots. The defualt is set to \code{NULL}.}

\item{ylim.rle.iqr.plot}{Numeric. Specifies the ylim of the boxplot or scatter plots when the RLE IQRs are used. If
\code{NULL}, the function will automatically determine an appropriate ylim for the plots. The default is set to \code{NULL}.}

\item{points.size}{Numeric. A numeric vlaue that specifies the point size of the scatter plots. The default is set to 1.}

\item{plot.ncol}{Numeric. A numeric value indicating the number of columns in the plot grid. When the number of selected
assays is more than 1, the function puts all the RLE boxplots in one grid. The default is set to 2.}

\item{plot.nrow}{Numeric. A numeric value indicating the number of rows in the plot grid. When the number of selected
assays is more than 1, the function puts all the RLE boxplots in one grid. The default is set to 3.}

\item{plot.output}{Logical. If \code{TRUE}, individual RLE plot(s) will be printed while the function is running.}

\item{save.se.obj}{Logical. Specifies whether to save the plots in the metadata of the SummarizedExperiment object or
to output them as a list. The default is set to \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE}, messages for different steps of the function will be shown.}
}
\value{
A SummarizedExperiment object that contains all the plot(s) in the metadata or a list containing all the plot(s).
}
\description{
This function plots a variable against the medians and IQR of a relative log expression (RLE) data. Because of the
sensitivity of the RLE medians and IQR to unwanted variation, we  examine the relationships between RLE medians and
IQR with potential sources of unwanted variation. In the absence of any influence of unwanted variation in the data,
we should see no such associations.
}
\details{
If the variable is categorical, boxplots of the RLE medians and IQR against the variable will be generated. If
the variable is continuous, scatter plots will be created.
}
\references{
Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
\author{
Ramyar Molania
}
