% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessVariablesForPrPs.R
\name{prepareVariableForPrPs}
\alias{prepareVariableForPrPs}
\title{Prepare variables for PRPS}
\usage{
prepareVariableForPrPs(
  se.obj,
  main.variable,
  other.variables = NULL,
  nb.mnn = 1,
  min.sample.for.ps = 3,
  clustering.method = "kmeans",
  nb.clusters = 3,
  other.uv.clustering.method = "kmeans",
  nb.other.uv.clusters = 3,
  select.extreme.groups = TRUE,
  cover.all.batches = FALSE,
  nb.batches.to.cover = 2,
  assess.variables.association = TRUE,
  plot.output = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{main.variable}{TTTT}

\item{other.variables}{TTTT}

\item{nb.mnn}{TTTT}

\item{min.sample.for.ps}{TTTT}

\item{clustering.method}{TTTT}

\item{nb.clusters}{TTTT}

\item{other.uv.clustering.method}{TTTT}

\item{nb.other.uv.clusters}{TTTT}

\item{select.extreme.groups}{TTTT}

\item{cover.all.batches}{TTTT}

\item{nb.batches.to.cover}{TTTT}

\item{assess.variables.association}{TTTT}

\item{plot.output}{TTTT}

\item{verbose}{Logical. Indicates whether to display output messages during function execution. The default is set to
\code{TRUE}.}
}
\description{
Prepare variables for PRPS
}
\author{
Ramyar Molania
}
