% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessData.R
\name{preProcessData}
\alias{preProcessData}
\title{Pre-process RNA-seq data.}
\usage{
preProcessData(
  se.obj,
  assay.name,
  normalization = "CPM",
  regress.out.variables = NULL,
  regress.out.rle.med = FALSE,
  apply.log = TRUE,
  pseudo.count = 1,
  check.se.obj = TRUE,
  remove.na = "assay",
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{description}

\item{assay.name}{description}

\item{normalization}{description}

\item{regress.out.variables}{description}

\item{regress.out.rle.med}{description}

\item{apply.log}{description}

\item{pseudo.count}{description}

\item{check.se.obj}{description}

\item{remove.na}{description}

\item{verbose}{description}
}
\description{
This function applies ....
}
\author{
Ramyar Molania
}
