% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressOutVariables.R
\name{regressOutVariables}
\alias{regressOutVariables}
\title{Regress out variables from RNA-seq data}
\usage{
regressOutVariables(
  se.obj,
  assay.name,
  variables,
  apply.log = TRUE,
  pseudo.count = 1,
  check.se.obj = TRUE,
  remove.na = "none",
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{description}

\item{assay.name}{description}

\item{variables}{description}

\item{apply.log}{description}

\item{pseudo.count}{TTTT}

\item{check.se.obj}{description}

\item{remove.na}{description}

\item{verbose}{description}
}
\description{
This functions uses a linear model to regresses out any specified variables from data in SummarizedExperiment object.
}
\author{
Ramyar Molania
}
