% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeAssays.R
\name{removeAssays}
\alias{removeAssays}
\title{Removes datasets (assays) from a SummarizedExperiment object}
\usage{
removeAssays(se.obj, assays.to.remove)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assays.to.remove}{Character. A character to a vector of characters that specifies the name of data (assays) that
will be removed from the SummarizedExperiment object.}
}
\description{
Removes datasets (assays) from a SummarizedExperiment object
}
