% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameAssays.R
\name{renameAssays}
\alias{renameAssays}
\title{Removes datasets (assays) from a SummarizedExperiment object}
\usage{
renameAssays(se.obj, new.names)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{new.names}{Character. A character to a vector of characters that specifies the name of data (assays) that
will be removed from the SummarizedExperiment object.}
}
\description{
Removes datasets (assays) from a SummarizedExperiment object
}
