% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameVariables.R
\name{renameVariables}
\alias{renameVariables}
\title{Renames variables in SummarizedExperiment}
\usage{
renameVariables(se.obj, current.names, new.names, keep.old.names = FALSE)
}
\arguments{
\item{se.obj}{description}

\item{current.names}{description}

\item{new.names}{description}

\item{keep.old.names}{description}
}
\description{
Renames variables in SummarizedExperiment
}
