% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectColors.R
\name{selectColor}
\alias{selectColor}
\title{Print colored messages.}
\usage{
selectColor(nb.color, group = "pan.selection")
}
\arguments{
\item{nb.color}{Message to be printed.}

\item{group}{The color of the message.}
}
\description{
Print colored messages.
}
