% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectColors.R
\name{selectColors}
\alias{selectColors}
\title{Print colored messages.}
\usage{
selectColors(nb.color, group = "pan.selection.a")
}
\arguments{
\item{nb.color}{Message to be printed.}

\item{group}{The color of the message.}
}
\description{
Print colored messages.
}
