% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectGeneSets.R
\name{selectGenesSets}
\alias{selectGenesSets}
\title{Selects genes for partial correlation and gene scoring analysis.}
\usage{
selectGenesSets(se.obj, gene.set = "immune.stromal", set.name = NULL)
}
\arguments{
\item{se.obj}{TTT}

\item{gene.set}{TTT}

\item{set.name}{TTT}
}
\description{
This function adds the TCGA RNA-seq batch information to the a SummarizedExperiment object.
}
\author{
Ramyar Molania
}
