% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga_batch_info.R
\docType{data}
\name{tcga.batch.info}
\alias{tcga.batch.info}
\title{TCGA Batch Information}
\format{
A data frame with X rows and Y variables:
\describe{
\item{Sample.ids.full.a}{Unique identifier for each sample}
\item{Sample.ids.full}{Batch information for each sample}
\item{Sample.ids.3}{Other metadata columns...}
\item{Sample.ids.4}{Other metadata columns...}
\item{Sample.ids.5}{Other metadata columns...}
\item{Sample.ids.6}{Other metadata columns...}
\item{Sample.ids.3.a}{Other metadata columns...}
\item{Sample.ids.4.a}{Other metadata columns...}
\item{Sample.ids.5.a}{Other metadata columns...}
\item{Sample.ids.6.a}{Other metadata columns...}
\item{Year}{Other metadata columns...}
\item{Month}{Other metadata columns...}
\item{Day}{Other metadata columns...}
\item{Plates}{Other metadata columns...}
\item{TSS}{Other metadata columns...}
\item{Center}{Other metadata columns...}
\item{Tissues}{Other metadata columns...}
\item{Cancer.type.a}{Other metadata columns...}
\item{Cancer.type}{Other metadata columns...}
}
}
\source{
\url{https://www.cancer.gov/tcga}
}
\usage{
tcga.batch.info
}
\description{
This dataset contains batch metadata for TCGA samples used in the RUVprps package.
}
\keyword{datasets}
