% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNCGs.R
\name{addNCGs}
\alias{addNCGs}
\title{Add pre-selected sets of NCGs to SummarizedExperiment object.}
\usage{
addNCGs(se.obj, ncg, subset.name = NULL, verbose = TRUE)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{ncg}{A logical value or a vector of gene names or IDs representing pre-selected NCGs. If gene names or IDs are
provided, the row names of the SummarizedExperiment object must match these names or IDs.}

\item{subset.name}{Character. A character that specifies the name of the NCG set in the metadata of the SummarizedExperiment
object. The default is set to \code{NULL}, in which case the function will generate a name as following:
\code{paste0(sum(ncg), '_psg')}.}

\item{verbose}{Logical. Indicates whether to display output messages during function execution. The default is set to
\code{TRUE}.}
}
\value{
A SummarizedExperiment object with a metadata that contains the NCGs.
}
\description{
This function adds pre-selected set(s) of gene(s) as negative control genes (NCGs) to a SummarizedExperiment object.
}
\details{
A pre-selected set of negative control genes (NCGs) will be stored in the following location:
se.obj->metadata->NCG->pre.selected->subset.name. These genes can be used for various analyses, including identifying
unknown sources of variation, assessing variation, performing RUV normalization, and evaluating normalization steps. The
gene set will be stored in: metadata->NCG->pre.selected->subset.name->gene.set
}
\author{
Ramyar Molania
}
