% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNCGs.R
\name{addNCGs}
\alias{addNCGs}
\title{Add pre-selected set of negative control genes to SummarizedExperiment object.}
\usage{
addNCGs(se.obj, ncg, save.se.obj = TRUE, output.name = NULL, verbose = TRUE)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{ncg}{A logical or vector of gene names of pre-selected genes as NCGs.}

\item{save.se.obj}{Symbol. Indicates}

\item{output.name}{Symbol. Indicates the name of the output in the meta data of the SummarizedExperiment object. The
default is 'NULL'. This means the function will create a name based the specified argument.}

\item{verbose}{Logical. If TRUE, displaying process messages is enabled.}
}
\description{
This function adds a pre-selected set of negative control genes to SummarizedExperiment object.
}
\author{
Ramyar Molania
}
