% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTcgaClinicalInfo.R
\name{addTcgaClinicalInfo}
\alias{addTcgaClinicalInfo}
\title{Adds the TCGA RNA-seq batch information}
\usage{
addTcgaClinicalInfo(
  se.obj,
  tissue.type,
  factors = c("OS", "OS.time"),
  cancer.type = NULL,
  missing.samples.name = "DF",
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{TTT}

\item{tissue.type}{TTT}

\item{factors}{TTTT}

\item{cancer.type}{TTTT}

\item{missing.samples.name}{TTTTT}

\item{verbose}{TTTTT}
}
\description{
This function adds the TCGA RNA-seq batch information to the a SummarizedExperiment object.
}
\author{
Ramyar Molania
}
