% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyLog.R
\name{applyLog}
\alias{applyLog}
\title{Apply log2 with a pseudo count on dataset(s) in SummarizedExperiment object.}
\usage{
applyLog(
  se.obj,
  assay.names = "all",
  pseudo.count = 1,
  check.se.obj = TRUE,
  remove.na = "assays",
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Character. A character string or a vector of characters strings specifying the name(s) of the data
(assays) in the SummarizedExperiment object to be selected to apply a log transformation. These dataset(s) will be
log2-transformed with a pseudo count. The default is se to \code{all}, which indicates that all data sets in the
SummarizedExperiment object will be selected.}

\item{pseudo.count}{Numeric. A numeric value as pseudo count value to be added to all measurements in the selected
dataset(s) before applying the log2 transformation, to avoid \code{-Inf} values for zero measurements. The default is set to 1.}

\item{check.se.obj}{Logical. Indicates whether to assess the structure fo the SummarizedExperiment object or not. The
default it is set to \code{TRUE}.}

\item{remove.na}{Character. A Character string that indicates whether to remove NA or missing values from the data sets or
not. The options are: \code{assays} or \code{none}. The default is set to \code{assays}.  Refer to the \code{checkSeObj()} function for more
details.}

\item{verbose}{Logical. Indicates whether to display output messages during function execution. The default is set to
\code{TRUE}.}
}
\value{
The function returns a log2 transformed of all specified data sets as a list object.
}
\description{
Apply log2 with a pseudo count on dataset(s) in SummarizedExperiment object.
}
\author{
Ramyar Molania
}
