% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessNCGs.R
\name{assessNCGs}
\alias{assessNCGs}
\title{Assess performance of NCG.}
\usage{
assessNCGs(
  se.obj,
  assay.name,
  variables.to.assess.ncg,
  ncg,
  apply.log = TRUE,
  pseudo.count = 1,
  nb.pcs,
  svd.bsparam = bsparam(),
  center = TRUE,
  scale = FALSE,
  plot.output = TRUE,
  check.se.obj = TRUE,
  remove.na = "none",
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{Character. A character or a vector of characters specifying the name(s) of the data (assays) in the
SummarizedExperiment object to be selected. These data stes will be log2-transformed with a pseudo count. The default
is se to \code{all}, which indicates that all data sets in the SummarizedExperiment object will be selected.}

\item{variables.to.assess.ncg}{TTT}

\item{ncg}{TTTT}

\item{apply.log}{TTTT}

\item{pseudo.count}{Numeric. A numeric value as pseudo count value to be added to all measurements in the selected
data sets before applying the log2 transformation, to avoid \code{-Inf} values for zero measurements. The default is set to 1.}

\item{nb.pcs}{TTT}

\item{svd.bsparam}{TTT}

\item{center}{TTTT}

\item{scale}{TTT}

\item{plot.output}{TTT}

\item{check.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object or not. The default it is
set to \code{TRUE}.}

\item{remove.na}{Character. A Character that indicates whether to remove NA or missing values from the data sets or
not. The options are: \code{assays} or \code{none}. The default is set to \code{assays}.  Refer to the \code{checkSeObj()} function for more
details.}

\item{verbose}{Logical. Indicates whether to display output messages during function execution. The default is set to
\code{TRUE}.}
}
\value{
The function returns a log2 transformed of all specified data sets as a list object.
}
\description{
Assess performance of NCG.
}
\author{
Ramyar Molania
}
