% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessNCGs.R
\name{assessNCGs}
\alias{assessNCGs}
\title{Assesses and compares the performance of NCGs gene sets.}
\usage{
assessNCGs(
  se.obj,
  assay.name,
  variables,
  bio.variables = NULL,
  uv.variables = NULL,
  ncg.type,
  ncg.group.name,
  ncg.set.names = "all",
  nb.pcs = 5,
  center = TRUE,
  scale = FALSE,
  apply.log = TRUE,
  pseudo.count = 1,
  plot.output = TRUE,
  output.name = NULL,
  check.se.obj = TRUE,
  remove.na = "none",
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{Character. A character that specifies the name of the data in the SummarizedExperiment object. This
data should one that will be used as the input data for RUV-III-PRPS normalization.}

\item{variables}{Character. One or more variable names in the SummarizedExperiment object. These will be used to assess
the performance of the NCGs. If this is set to \code{NULL}, both \code{bio.variables} and \code{uv.variables} must be specified.}

\item{bio.variables}{Character. One or more variable names representing biological variables (e.g., cancer subtypes,
tumor purity) within the SummarizedExperiment object. This can be a vector of categorical, continuous, or mixed variable
types. The default is set to \code{NULL}. This must be specified if the performance scores is need to be calculated.}

\item{uv.variables}{Character. One or more variable names representing unwanted variables (e.g., batch effects,
library size) within the SummarizedExperiment object. This can also be a vector of categorical, continuous, or mixed
variable types. The default is set to \code{NULL}. This must be specified if the performance scores is need to be calculated.}

\item{ncg.type}{Character. A character that specifies the type of NCGs to be selected and assessed their performnce.
Options are \code{'supervised'}, \code{'un.supervised'}, \code{pre.selected} and \code{all.genes}.}

\item{ncg.group.name}{Character. A character that specifies the name of the group of NCGs stored under the 'supervised' or
'un.supervised' slots in the SummarizedExperiment object. Refer to supervised and unsupervised functions for finding
NCGS for more details.}

\item{ncg.set.names}{Character. Specifies the exact name of an NCG set under the given \code{ ncg.group.name} in either the
'supervised' or 'un.supervised' slot or in the \code{pre.selected} slot. The default is se to \code{'all'}, which selects all
available NCG sets.}

\item{nb.pcs}{Numeric. A numeric value that indicates the number of principal components (PCs) to compute and use for
the assessment analysis. The default is set to \code{10}.}

\item{center}{Logical. Indicates whether to center the data prior to applying PCA suing SVD. If \code{TRUE}, centering is
done by subtracting the column means from the respective columns. The default is set to \code{TRUE}.}

\item{scale}{Logical. Indicates whether to scale the data prior to applying PCA using SVD. If \code{TRUE}, the centered
columns are divided by their standard deviations (if \code{center = TRUE}), or by the root mean square
(if \code{center = FALSE}). The default is set to \code{FALSE}.}

\item{apply.log}{Logical. Indicates whether to apply a log transformation to the data before computing the SVD. The
default is set to \code{TRUE}. Data should be in log scale before applying SVD.}

\item{pseudo.count}{Numeric. A numeric value as a pseudo count to be added to all measurements before applying the
log transformation. The default is set to \code{1}.}

\item{plot.output}{Logical. Indicates whether to generate the assessment plot during function execution. The default
is set to \code{TRUE}.}

\item{output.name}{Character. Specifies the name under which the assessment plot will be saved in the
SummarizedExperiment object. If \code{NULL}, a name will be generated using:
\code{paste0('comparison_', paste0(names(ncgs), collapse = '&'))}.}

\item{check.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object or not before applying
nationalizations. The default is set  to \code{TRUE}. We refer to the \code{checkSeObj()} function for more details.}

\item{remove.na}{Character. A character that indicates whether to remove NA or missing values from the data sets or
not. The options are \code{'assays'} or \code{'none'}. The default is set to \code{'assays'}. Refer to the \code{checkSeObj()} function
for more details.}

\item{save.se.obj}{Logical. Indicates whether to save the assessment plot metadata of the SummarizedExperiment object.
The default is set to \code{TRUE}. The plot will be saved to:}

\item{verbose}{Logical. Indicates whether to display output messages during function execution. The default is set
to \code{TRUE}.}
}
\value{
Either a SummarizedExperiment object containing the NCG assessment plots, or out theses results as list if
\code{save.se.obj} is set to \code{FALSE}.
}
\description{
This function globally assesses and compares the performance of various sets of negative control genes (NCGs) based on
their ability to capture unwanted variation while remaining unassociated with biological variation. This is essential
to select a suitable set of NCGs for RUV-III normalization.
}
\details{
For each NCG set, the function performs PCA on the specified data using only the genes in the set. It then evaluates
the association of the top \code{nb.pcs} principal components with variables as defined by \code{variables}. Linear regression
is used for continuous variables, and vector correlation is used for categorical variables. For each NCG
set, $R^2$ values (from regression) and vector correlation coefficients are computed to quantify these associations.
The $R^2$ and vector correlation values calculated for individual variables are plotted against the cumulative PCs.
Ideally, a suitable set of NCGs should show high and low correlation values, respectively, for the unwanted and biological
variables.\
If both \code{bio.variables} and \code{ub.variables} are specified, then the function generates a performance score between 0 and
1 is for NCG each set, reflecting how effectively the NCGs capture unwanted variation while minimizing association
with biological factors. To compute such scores, the correlations obtained for biological variables are subtracted
from 1, and then the correlations for biological and unwanted variables are averaged, separately. The final score for
each NCG set is calculated as the sum of half the average scores. A higher score indicates a better set of NCGs.
}
\author{
Ramyar Molania
}
