% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessNormalization.R
\name{assessNormalization}
\alias{assessNormalization}
\title{Assesses the performance of normalization methods.}
\usage{
assessNormalization(
  se.obj,
  assay.names = "all",
  bio.variables,
  uv.variables,
  assessments.to.exclude = NULL,
  fast.pca = TRUE,
  sil.dist.measure = "euclidian",
  sli.nb.pcs = 3,
  ari.clustering.method = "hclust",
  ari.hclust.method = "complete",
  ari.hclust.dist.measure = "euclidian",
  ari.nb.pcs = 3,
  corr.method = "spearman",
  corr.cutoff = 0.2,
  anova.method = "aov",
  fvalue.cutoff = 1,
  vca.nb.pcs = 3,
  lra.nb.pcs = 3,
  pcorr.method = "spearman",
  pcorr.cutoff = 0.3,
  bio.weight = 0.6,
  uv.weight = 0.4,
  plot.output = TRUE,
  save.se.obj = TRUE,
  output.name = NULL,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{character or character vector. One or more names of assays to select from the SummarizedExperiment
object.
The default is 'all', which selects all assays in the object.}

\item{bio.variables}{character or character vector. One or more column names indicating known biological variables
(categorical or continuous) in the SummarizedExperiment object.}

\item{uv.variables}{character or character vector. One or more column names indicating unwanted variables
(categorical or continuous) in the SummarizedExperiment object.}

\item{assessments.to.exclude}{character or character vector. Names of assessment metrics to exclude, as returned by
the \code{getAssessmentMetrics()} function. Default is NULL.}

\item{fast.pca}{Logical. Whether to use fast PCA. Default is TRUE.}

\item{sil.dist.measure}{character. Distance measure to use for silhouette analysis. Options: 'euclidean', 'maximum',
'manhattan', 'canberra', 'binary', or 'minkowski'. Default is 'euclidean'.}

\item{sli.nb.pcs}{Numeric. Number of principal components to use for silhouette analysis. Default is 3. Must not exceed
\code{compute.nb.pcs}.}

\item{ari.clustering.method}{character. Clustering method for ARI computation. Options: 'mclust' or 'hclust'. Default
is 'hclust'.}

\item{ari.hclust.method}{character. Method used in \code{hclust} for ARI. Options: 'ward.D', 'ward.D2', 'single', 'complete',
'average', 'mcquitty', 'median', or 'centroid'. Default is 'complete'.}

\item{ari.hclust.dist.measure}{character. Distance measure for \code{hclust} in ARI analysis. Options:'euclidean', 'maximum',
'manhattan', 'canberra', 'binary', or 'minkowski'. Default is 'euclidean'.}

\item{ari.nb.pcs}{Numeric. Number of principal components to use for ARI analysis. Default is 3. Must not exceed
\code{compute.nb.pcs}.}

\item{corr.method}{character. Correlation method for gene-variable correlation. Options: 'spearman' or 'pearson'.
Default is 'spearman'.}

\item{corr.cutoff}{Numeric. Threshold for selecting genes with an absolute correlation lower than this value with
continuous variables. The default is set to 0.2.}

\item{anova.method}{character. ANOVA method for gene-variable comparison. Options: 'aov' or 'welch'. Default is 'aov'.}

\item{fvalue.cutoff}{Numeric. Threshold for selecting genes with an F-statistic lower than this value with categorical
variables. Te default is set to 1.}

\item{vca.nb.pcs}{Numeric. Number of principal components to use for variance component analysis. Default is 3. Must
not exceed \code{compute.nb.pcs}.}

\item{lra.nb.pcs}{Numeric. Number of principal components to use for latent representation analysis. Default is 3.
Must not exceed \code{compute.nb.pcs}.}

\item{pcorr.method}{character. Correlation method for gene-gene ordinary and partial correlations. Options: 'spearman'
or 'pearson'. The default is set to \code{spearman}.}

\item{pcorr.cutoff}{Numeric. Threshold for selecting gene-gene correlations where the difference between ordinary and
partial correlations is less than this value.}

\item{bio.weight}{Numeric. Weight for the biological preservation score. Default is 0.6. See details for more information.}

\item{uv.weight}{Numeric. Weight for the removal of unwanted variation score. Default is 0.4. See details for more
information.}

\item{plot.output}{Logical. If TRUE, displays the final assessment plot.}

\item{save.se.obj}{Logical. If TRUE, saves the results in the SummarizedExperiment object. Default is TRUE.}

\item{output.name}{character. File name for the results stored in the SummarizedExperiment. If NULL, the function creates
a default name using \code{paste0('RUVprps_assessNormalization', length(assays(se.obj)), '_assays.')}.}

\item{verbose}{Logical. If TRUE, displays messages during function execution.}
}
\value{
A SummarizedExperiment object containing the assessment matrix, plot, and table, or a list containing all results.
}
\description{
This function summarizes a range of global and gene level metrics obtained using the \code{assessVariation} function for
individual biological and unwanted variables. The functions returns numerical assessments in a table to assess the
performance of difference normalization. Refer to details for more information.
}
\details{
It is essential to assess the performance of a normalization as a remover of unwanted variation and a preserver of
biological variation in the data. An ideal normalization should preserve all known sources of biological variation
and reduce or remove the impact of all known sources of unwanted variation. To rank methods according to their
performance, we aggregate all evaluation metrics into a overall score as follows.
Several assessment will be performed:
For each categorical variable:
\itemize{
\item PCA plot of the categorical variable.
\item Silhouette and ARI computed on the categorical variable.
\item Differential analysis based ANOVA between the gene expression and the categorical variable.
\item Vector correlation between the first cumulative PCs of the gene expression and the categorical variable.
For each continous variable:
\item Linear regression between the first cumulative PC and continuous variable.
\item Correlation between gene expression and continuous variable.
}

It will output the following plots:
\itemize{
\item PCA plot of each categorical variable.
\item Boxplot of the F-test distribution from ANOVA between the gene expression and each categorical variable.
\item Vector correlation between the first cumulative PCs of the gene expression and each categorical variable.
\item Combined Silhouette plot of the combined pair of all categorical variables.
\item Linear regression between the first cumulative PC and continuous variable.
\item Boxplot of the correlation between gene expression and continuous variable.
\item It will also output the RLE plot distribution.
}
}
\references{
**Molania R., ..., Speed, T. P., A new normalization for Nanostring nCounter gene expression data, Nucleic Acids
Research, 2019.
**Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
\author{
Ramyar Molania
}
