% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeLmmFormula.R
\name{changeLmmFormula}
\alias{changeLmmFormula}
\title{Changes the formula for linear mixed model}
\usage{
changeLmmFormula(form, out.put = "character", sub.set = NULL)
}
\arguments{
\item{form}{Character. A character string specifying linear mixed model form.}

\item{out.put}{Character. A character string specifying the type of output of function. The options are \code{character}
and \code{sub.formula}. The default is set to \code{character}.}

\item{sub.set}{Character. A character string specifying which subset of the form should be extracted. The defulat is
set to \code{NULL}.}
}
\description{
Changes the formula for linear mixed model
}
