% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCCA.R
\name{computeCCA}
\alias{computeCCA}
\title{Find canonical correlation coordinates in RNA-seq data}
\usage{
computeCCA(
  se.obj,
  assay.name,
  variable,
  reference.group = NULL,
  nb.cca = 10,
  hvg = NULL,
  scale = TRUE,
  normalization = "CPM",
  regress.out.variables = NULL,
  regress.out.rle.med = FALSE,
  cosine.norm = FALSE,
  sample.to.use = "all",
  apply.log = TRUE,
  pseudo.count = 1,
  check.se.obj = TRUE,
  remove.na = "none",
  cca.set.name = NULL,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{TTTTT}

\item{variable}{TTTTT}

\item{reference.group}{TTTTT}

\item{nb.cca}{TTTTT}

\item{hvg}{TTTTT}

\item{scale}{TTTTT}

\item{normalization}{TTTTT}

\item{regress.out.variables}{TTTTT}

\item{regress.out.rle.med}{TTTTT}

\item{cosine.norm}{TTTTT}

\item{sample.to.use}{TTTTT}

\item{apply.log}{TTTTT}

\item{pseudo.count}{TTTTT}

\item{check.se.obj}{TTTTT}

\item{remove.na}{TTTTT}

\item{cca.set.name}{TTTTT}

\item{save.se.obj}{TTTTT}

\item{verbose}{TTTTT}
}
\description{
Find canonical correlation coordinates in RNA-seq data
}
\author{
Ramyar Molania
}
