% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGenesPartialCorrelation.R
\name{computeGenesPartialCorrelation}
\alias{computeGenesPartialCorrelation}
\title{Compute gene-gene partial correlation analysis.}
\usage{
computeGenesPartialCorrelation(
  se.obj,
  assay.names,
  variable,
  method = "spearman",
  genes = NULL,
  select.genes = TRUE,
  reference.data = NULL,
  corr.coff.cutoff = 0.3,
  apply.log = TRUE,
  pseudo.count = 1,
  apply.round = TRUE,
  check.se.obj = TRUE,
  remove.na = "none",
  override.check = FALSE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Character. A character string or vector of character strings specifying the name(s) of the data set(s)
in the SummarizedExperiment object for calculating the correlations. The default is set to \code{all}, which indicates that
all assays of the SummarizedExperiment object will be selected.}

\item{variable}{Character. A character string indicating the column name in the SummarizedExperiment object that
contains a continuous variable, such as library size, tumor purity, etc.}

\item{method}{Character. Specifies which correlation method should be used. The options are \code{pearson}, \code{kendall}, or
\code{spearman}. The default is set to \code{spearman}.}

\item{genes}{Character or Numeric. A vector containing genes on which the correlation analysis will be performed. The
vector can be logical, numeric, or gene names (symbol). The default is to \code{NULL}, which means all genes will be selected.}

\item{select.genes}{Logical. If \code{TRUE}, the function will compute the correlation between individual genes and the
specified variable, then select a subset of genes based on the "corr.coff.cutoff" for downstream analysis. The default
is set to \code{TRUE}. This will speed up the computational time.}

\item{reference.data}{Character. A character string specifying the name of the data to be used for selecting genes
based on the correlation analysis. The default is se to \code{NULL}, which means all specified assays will be used.}

\item{corr.coff.cutoff}{Numeric. A numeric value used as a cutoff for selecting genes. The default is set to 0.7.}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data before computing the correlation.
The default is se to \code{TRUE}.}

\item{pseudo.count}{Numeric. A numeric value representing a pseudo count to be added to all measurements of the data
set(s) before applying the log transformation to avoid \code{-Inf} for measurements that are equal to 0. The default is set
to 1.}

\item{apply.round}{Logical. Indicates whether to round the correlation coefficients. The default is set to \code{TRUE}.}

\item{check.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object. See the \code{checkSeObj()}
function for more details.}

\item{remove.na}{Character. Specifies whether to remove NA or missing values from the assays. The options are \code{assays}
and \code{none}. The default is set to \code{assays}, meaning that all NA or missing values from the assays will be removed
before computing RLE. See the \code{checkSeObj()} function for more details.}

\item{override.check}{Logical. When set to TRUE, the function verifies whether the PPcorr has already been computed
for the current parameters on the SummarizedExperiment object. If it has, the metric will not be recalculated. The
default is set to  \code{FALSE}.}

\item{save.se.obj}{Logical. Indicates whether to save the correlation results in the metadata of the SummarizedExperiment
object or to output the result as a list. The default is set to \code{TRUE}.}

\item{verbose}{Logical. If TRUE, displays the messages for different steps of the function.}
}
\value{
A SummarizedExperiment object containing the correlation coefficients or a list of the correlation coefficients
for individual assays.
}
\description{
This function computes all possible gene-gene pairwise ordinary and partial correlation of the dataset(s) in the
SummarizedExperiment object.

Molania R., et al. (2023). Removing unwanted variation from large-scale RNA sequencing data with PRPS. \emph{Nature Biotechnology}
\link{https://doi.org/10.1038/s41587-022-01440-w}
}
\details{
Partial correlation estimates the correlation between two variables while controlling for the effect of a third variable.
In the context of RNA-seq data, this allows computation of gene–gene correlations while accounting for confounding factors
such as library size or tumor purity.
The function \code{computeGenesPartialCorrelation()} calculates both ordinary
and partial correlations for all possible gene–gene pairs. The formula for the
partial correlation \( r_{xy.z} \) is:
\deqn{r_{xy.z} = \frac{r_{xy} - r_{xz} \cdot r_{yz}}{\sqrt{(1 - r_{xz}^2) \cdot (1 - r_{yz}^2)}}}

where:
\itemize{
\item \( r_{xy} \): correlation between variables \( x \) and \( y \)
\item \( r_{xz} \): correlation between \( x \) and the control variable \( z \)
\item \( r_{yz} \): correlation between \( y \) and the control variable \( z \)
}

The companion function \code{plotGenesPartialCorrelation()} generates histograms,
scatter plots, and boxplots of both ordinary and partial correlation coefficients
for individual assays.
}
\note{
Computing partial correlations for all gene pairs in large RNA-seq datasets
can be computationally intensive. It is recommended to restrict the analysis
to a subset of genes that show strong association with the variable of interest.
}
\author{
Ramyar Molania
}
