% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGenesVarianceComposition.R
\name{computeGenesVarianceComposition}
\alias{computeGenesVarianceComposition}
\title{Computes gene-level variance composition across variables in RNA-seq.}
\usage{
computeGenesVarianceComposition(
  se.obj,
  assay.names,
  form,
  adjust.data = FALSE,
  adjustment.form = NULL,
  adjustment.method = "lmm",
  samples.to.use = "all",
  apply.log = TRUE,
  pseudo.count = 1,
  nb.cores = 10,
  plot.output = TRUE,
  output.name = NULL,
  check.se.obj = TRUE,
  remove.na = "both",
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Character. A character string or vector of character strings specifying the name(s) of the assay(s)
in the SummarizedExperiment object to compute the gene variance composition By default is set to 'all', which means all
assays of the SummarizedExperiment object will be selected.}

\item{form}{TTT}

\item{adjust.data}{TTT}

\item{adjustment.form}{TTT}

\item{adjustment.method}{TTT}

\item{samples.to.use}{TTT}

\item{apply.log}{Logical. Indicates whether to apply a log-transformation to the data before performing ANOVA. The
default is set to \code{TRUE}.}

\item{pseudo.count}{Numeric. A numeric value representing a pseudo count to be added to all measurements before applying
the log transformation. The default is set to 1.}

\item{nb.cores}{TTTT}

\item{plot.output}{TTTT}

\item{output.name}{TTT}

\item{check.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object. The default is set to \code{TRUE}.
This means the function will apply the \code{checkSeObj()} function.}

\item{remove.na}{Character. A character string specifying whether to eliminate missing values from \code{assays}, \code{sample.annotation},
\code{both}, or \code{none}. When 'assays' is chosen, genes with missing values will be omitted. If 'sample.annotation' is selected,
samples with NA or missing values for each 'variable' will be excluded. The default is 'both'.}

\item{save.se.obj}{Logical. Indicates whether to save the results, ANOVA F-statistics, and p-values in the metadata
of the SummarizedExperiment object or to output these results as a list or vector. The default is set to \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE}, displays the messages of different steps of the function.}
}
\value{
Either a SummarizedExperiment object containing the log2 F-statistics and p-values of ANOVA for the continuous
variable or a list of these results.
}
\description{
TTTT
}
\author{
Ramyar Molania
}
