% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countNCG.R
\name{countNCG}
\alias{countNCG}
\title{Count the current NCG sets in SummarizedExperiment object.}
\usage{
countNCG(
  se.obj,
  ncg.selection,
  ncg.group.name,
  create.venn.diagram = FALSE,
  check.all = FALSE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{ncg.selection}{TTTT}

\item{ncg.group.name}{TTTT}

\item{create.venn.diagram}{TTTT}

\item{check.all}{TTT}
}
\value{
A SummarizedExperiment object with a metadata that contains the NCGs.
}
\description{
This function adds pre-selected sets of negative control genes (NCGs) to a SummarizedExperiment object.
}
\details{
A pre-selected set of negative control genes (NCGs) will be stored in the following location:
se.obj->metadata->NCG->pre.selected->subset.name. These genes can be used for various analyses, including identifying
unknown sources of variation, assessing variation, performing RUV normalization, and evaluating normalization steps. The
gene set will be stored in: metadata->NCG->pre.selected->subset.name->gene.set
}
\author{
Ramyar Molania
}
