% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPrPsUnSupervised.R
\name{createPrPsUnSupervised}
\alias{createPrPsUnSupervised}
\title{Creates unsupervised pseudo-replicates of pseudo samples (PRPS).}
\usage{
createPrPsUnSupervised(
  se.obj,
  assay.name,
  uv.variables,
  approach = "mnn",
  data.input = "expr",
  clustering.method = "kmeans",
  nb.clusters = 3,
  other.uv.variables = NULL,
  other.uv.clustering.method = "kmeans",
  nb.other.uv.clusters = 3,
  hvg = NULL,
  select.extreme.groups = FALSE,
  min.batches.to.cover = "all",
  filter.prps.sets = TRUE,
  max.prps.sets = 3,
  min.sample.for.ps = 3,
  max.sample.for.ps = 10,
  nb.mnn = 1,
  nb.knn = 2,
  anchor.features = 2000,
  sct.clip.range = NULL,
  reduction = "cca",
  l2.norm = TRUE,
  dims = 1:10,
  k.anchor = 3,
  k.filter = 20,
  k.score = 5,
  max.features = 200,
  nn.method = "annoy",
  n.trees = 50,
  eps = 0,
  nb.pcs = 2,
  center = TRUE,
  scale = FALSE,
  svd.bsparam = bsparam(),
  normalization = "CPM",
  apply.cosine.norm = FALSE,
  regress.out.variables = NULL,
  check.prps.connectedness = FALSE,
  apply.log = TRUE,
  pseudo.count = 1,
  mnn.bpparam = SerialParam(),
  mnn.nbparam = KmknnParam(),
  assess.se.obj = TRUE,
  remove.na = "both",
  prps.sets.name = NULL,
  prps.group = NULL,
  plot.output = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{Character. A character string indicating the assay name within the SummarizedExperiment object for
the creation of PRPS data. The assay must be the one that will be used as data input for the RUV-III-PRPS normalization.}

\item{uv.variables}{Character. A character string or vector of strings specifying the name(s) of column(s) in the sample annotation
of the SummarizedExperiment object. This variable can be categorical or continuous. If a continuous variable
is provided, this will be divided into groups using the clustering method.}

\item{approach}{Character. Indicates which method to be used. The options are \code{anchor} and \code{mnn}.}

\item{data.input}{Character. Indicates which data should be used as input for finding the k nearest neighbors. Options
include: \code{expr} and \code{pcs}. If \code{pcs} is selected, the first PCs of the data will be used as input. If \code{expr} is selected,
the data will be selected as input.}

\item{clustering.method}{Character. A character string indicating the choice of clustering method for grouping the \code{uv.variable}
if a continuous variable is provided. Options include \code{kmeans}, \code{cut}, and \code{quantile}. The default is set to \code{kmeans}.}

\item{nb.clusters}{Numeric. A numeric value indicating how many clusters should be found if the \code{uv.variable} is a
continuous variable. The default is 3.}

\item{other.uv.variables}{TTTT}

\item{other.uv.clustering.method}{TTT}

\item{nb.other.uv.clusters}{TTT}

\item{hvg}{Vector. A vector containing the names of highly variable genes. These genes will be utilized to identify
anchor samples across different batches. The default value is set to \code{NULL}}

\item{select.extreme.groups}{TTTT}

\item{min.batches.to.cover}{TTTT}

\item{filter.prps.sets}{TTTT}

\item{max.prps.sets}{Numeric. The maximum number of PRPS sets across batches. The default is 10.}

\item{min.sample.for.ps}{Numeric. The minimum number of samples to be averaged to create a pseudo-sample. The default
is 3. The minimum value is 2.}

\item{max.sample.for.ps}{Numeric value indicating the maximum number of samples to be averaged for creating a pseudo-sample.
The default is \code{inf}. Please note that averaging a high number of samples may lead to inaccurate PRPS estimates.}

\item{nb.mnn}{Numeric. The maximum number of mutual nearest neighbors to compute. The default is 3.}

\item{nb.knn}{Numeric. The maximum number of nearest neighbors to compute. The default is 3.}

\item{anchor.features}{Numeric. A numeric value indicating the number of features to be used in anchor finding.
The default is 2000. We refer to the \code{FindIntegrationAnchors} R function for more details.}

\item{sct.clip.range}{Numeric. A numeric vector of length two specifying the min and max values the Pearson residual will be
clipped to. The default is \code{NULL}. See \code{FindIntegrationAnchors} for more.}

\item{reduction}{Character. Indicates which dimensional reduction to perform when finding anchors. Options are \code{cca},
\code{rpca}, and \code{rlsi}. The default is \code{cca}.}

\item{l2.norm}{Logical. Indicates whether to perform L2 normalization on the CCA sample embeddings after dimensional
reduction. Default is \code{TRUE}.}

\item{dims}{Numeric. Indicates which dimensions to use from CCA to define the neighbor search space. Default is 10.}

\item{k.anchor}{Numeric. Number of neighbors (k) used to select anchors. Default is 3.}

\item{k.filter}{Numeric. Number of neighbors (k) used for anchor filtering. Default is 200.}

\item{k.score}{Numeric. Number of neighbors (k) used when scoring anchors. Default is 30.}

\item{max.features}{Numeric. Maximum number of features used for neighborhood search space in anchor filtering. Default is 30.}

\item{nn.method}{Character. Method for nearest neighbor finding. Options: \code{rann}, \code{annoy}. Default is \code{annoy}.}

\item{n.trees}{Numeric. Number of trees for \code{annoy} search (more trees = higher precision).}

\item{eps}{Numeric. Error bound on the neighbor finding algorithm (from RANN or Annoy).}

\item{nb.pcs}{Numeric. Number of principal components to use as input for nearest neighbor search. Must be set if
\code{data.input = pcs}. Default is 2.}

\item{center}{Logical. Whether to center the data (subtract column means). Default is \code{TRUE}.}

\item{scale}{Logical. Whether to scale the data before applying SVD. Default is \code{FALSE}.}

\item{svd.bsparam}{A \code{BiocParallelParam} object defining parallelization for SVD. Default: \code{bsparam()}. See \code{runSVD}
from the BiocSingular package.}

\item{normalization}{Character. Normalization method before computing nearest neighbors. Default: \code{cpm}. If \code{NULL},
no normalization is applied.}

\item{apply.cosine.norm}{TTTT}

\item{regress.out.variables}{Character. Column names containing biological variables to regress out before estimating
unwanted variation. Default: \code{NULL}.}

\item{check.prps.connectedness}{TTT}

\item{apply.log}{Logical. Whether to apply a log transformation. Default is \code{TRUE}.}

\item{pseudo.count}{Numeric. Value added to counts before log transform to avoid \code{-Inf}. Default is 1.}

\item{mnn.bpparam}{Character. A \code{BiocParallelParam} object for parallel MNN search. Default: \code{SerialParam()}. See \code{findMutualNN}.}

\item{mnn.nbparam}{Character. A \code{BiocParallelParam} object for nearest neighbor search. Default: \code{KmknnParam()}.}

\item{assess.se.obj}{Logical. Whether to assess the SummarizedExperiment object using \code{checkSeObj}.}

\item{remove.na}{Character. Indicates whether to remove \code{NA} values. Options: \code{assays}, \code{none}. Default is \code{assays}.}

\item{prps.group}{Character.}

\item{plot.output}{TTTT}

\item{save.se.obj}{Logical. If \code{TRUE}, saves results in \code{metadata} of the SummarizedExperiment object. Otherwise,
returns results as a list. Default is \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE}, prints messages throughout function execution.}

\item{output.name}{Character.}
}
\description{
This function uses the \code{createUnSupervisedPRPSbyAnchors} or \code{createUnSupervisedPRPSbyMNN} to create PRPS sets of all
the unwanted variables in situations when the biological variation are unknown.
}
\author{
Ramyar Molania
}
