% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indentifyUnknownUV.R
\name{identifyUnknownUV}
\alias{identifyUnknownUV}
\title{Identifies potential unknown sources of unwanted variation in RNA-seq data.}
\usage{
identifyUnknownUV(
  se.obj,
  assay.name,
  approach = "rle",
  rle.comp = "median",
  regress.out.bio.variables = NULL,
  regress.out.bio.gene.sets = NULL,
  uv.gene.sets = NULL,
  ncg = NULL,
  clustering.methods = "nbClust",
  nbClust.diss = NULL,
  nbClust.distance = "euclidean",
  nbClust.min.nc = 2,
  nbClust.max.nc = 5,
  nbClust.method = "kmeans",
  nbClust.index = "silhouette",
  nbClust.alphaBeale = 0.1,
  max.samples.per.batch = NULL,
  nb.clusters = 3,
  apply.log = TRUE,
  pseudo.count = 1,
  nb.pcs = 2,
  center = TRUE,
  scale = FALSE,
  svd.bsparam = bsparam(),
  remove.current.estimates = FALSE,
  output.name = NULL,
  check.se.obj = TRUE,
  remove.na = "none",
  save.se.obj = TRUE,
  plot.output = TRUE,
  order.batches = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.name}{Character string. Specifies the assay name to be used for identifying potential sources of unwanted
variation.}

\item{approach}{Character string. One of \code{rle}, \code{pca}, or \code{sample.scoring}, specifying the method to detect unknown variation.}

\item{rle.comp}{Character string. One of \code{median}, \code{iqr}, or \code{both}. Specifies which RLE summary statistic to use.
The default is set to \code{median}.}

\item{regress.out.bio.variables}{Character string or vector. Column name(s) of biological variables in the sample
annotation.The default is set to \code{NULL}.}

\item{regress.out.bio.gene.sets}{List. Biological gene signatures to regress out before identifying unwanted variation.
The default is set to \code{NULL}.}

\item{uv.gene.sets}{List. Gene sets related to unwanted variation for use in \code{sample.scoring}.The default is set to \code{NULL}.}

\item{ncg}{Vector. Negative control genes. If not \code{NULL}, analysis will be restricted to these genes.The default is
set to \code{NULL}.}

\item{clustering.methods}{Character string. Clustering method: one of \code{kmeans}, \code{cut}, \code{quantile}, or \code{nbClust}.
The default is set to \code{nbClust}.}

\item{nbClust.diss}{Dissimilarity matrix. If provided, \code{nbClust.distance} must be \code{NULL}.The default is set to \code{NULL}.}

\item{nbClust.distance}{Character string. Distance measure: \code{euclidean}, \code{maximum}, \code{manhattan}, \code{canberra}, \code{binary},
\code{minkowski}, or \code{NULL}.The default is set to \code{euclidean}.}

\item{nbClust.min.nc}{Numeric. Minimum number of clusters. Must be between 1 and (n - 1).}

\item{nbClust.max.nc}{Numeric. Maximum number of clusters. Must be ≥ \code{nbClust.min.nc}.The default is set to 15.}

\item{nbClust.method}{Character string. Clustering method: \code{ward.D}, \code{ward.D2}, \code{single}, \code{complete}, \code{average},
\code{mcquitty}, \code{median}, \code{centroid}, or \code{kmeans}.}

\item{nbClust.index}{Character string. Clustering index to evaluate: e.g., \code{silhouette}, \code{gap}, \code{ch}, \code{db}, \code{all}, etc.}

\item{nbClust.alphaBeale}{Numeric. Significance threshold for Beale's index.}

\item{max.samples.per.batch}{Numeric. Max proportion of samples per cluster (only applies to \code{nbClust}).The default
is set to 0.1.}

\item{nb.clusters}{Numeric. Number of clusters when using \code{kmeans}, \code{cut}, or \code{quantile}.The default is set to 3.}

\item{apply.log}{Logical. Apply log-transformation before analysis.The default is set to \code{TRUE}.}

\item{pseudo.count}{Numeric. Pseudo count added before log-transforming data.The default is set to 1.}

\item{nb.pcs}{Numeric. Number of principal components to use in \code{pca} approach.The default is set to 2.}

\item{center}{Logical. Whether to center data before PCA.The default is set to \code{TRUE}.}

\item{scale}{Logical. Whether to scale data before PCA.The default is set to \code{FALSE}.}

\item{svd.bsparam}{A \code{BiocParallelParam} object. Controls parallelization for SVD computation.The default is set to
\code{bsparam()}.}

\item{remove.current.estimates}{Character string. Whether to remove current estimates of unknown batches.The default
is set to \code{TRUE}.}

\item{output.name}{Character string. Output file name. If \code{NULL}, a name is automatically generated.}

\item{check.se.obj}{Logical. If \code{TRUE}, the \code{checkSeobj} function is applied to validate the object.The default is set to
\code{TRUE}.}

\item{remove.na}{Character string. Where to remove \code{NA} values: \code{assays}, \code{sample.annotation}, \code{both}, or \code{none}.
The default is set to \code{both}.}

\item{save.se.obj}{Logical. If \code{TRUE}, results are saved in the metadata under \code{metadata$UV$Unknown}.The default is set to \code{TRUE}.}

\item{plot.output}{Logical. If \code{TRUE}, generates clustering input plots colored by identified groups.The default is set to \code{TRUE}.}

\item{order.batches}{Logical. If \code{TRUE}, orders estimated batches in the plot.The default is set to \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE}, displays messages throughout function execution.}
}
\description{
This function identifies unknown sources of unwanted variation in RNA-seq data using different robust statistical
approaches.
}
\details{
Identification of sources of unwanted variation is essential for creating PRPS data for RUV-III normalization.
There may be instances where certain sources of unwanted variation remain unrecorded; these are referred to as
"unknown" sources. This function uses three different approaches—\code{rle}, \code{pca}, and \code{sample.scoring}—to identify
potential unknown sources of unwanted variation when none are known.
\itemize{
\item In the \code{rle} approach, a clustering method specified by \code{clustering.methods} is applied to the RLE medians,
IQRs, or both. In the absence of unwanted variation, no distinguishable clusters should form.
\item In the \code{pca} approach, principal component analysis is applied to either a set of negative control genes or
all genes. The first PCs are then clustered to detect unknown sources of variation.
\item In the \code{sample.scoring} approach, samples are scored against gene sets (e.g., housekeeping genes) whose variation
may indicate unwanted variation. A clustering method is then applied to the scores.
}
}
\references{
\enumerate{
\item Gandolfo L. C. & Speed, T. P., RLE plots: visualizing unwanted variation in high dimensional data. PLoS ONE, 2018.
\item Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
}
\author{
Ramyar Molania
}
