% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDGE.R
\name{plotDGE}
\alias{plotDGE}
\title{Plot p-values histograms of DGE analysis.}
\usage{
plotDGE(
  se.obj,
  assay.names = "all",
  variable,
  method = "limma",
  plot.ncol = 1,
  plot.nrow = 2,
  plot.output = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Character or character vector. Specifies the name(s) of the assay(s) in the
SummarizedExperiment object to compute DGE. The default is set to \code{all}, indicating all assays of the
SummarizedExperiment object will be selected.}

\item{variable}{Character. Specifies the column name in the SummarizedExperiment object containing a categorical variable,
such as sample types or batches.}

\item{method}{Character. Specifies the differential expression method to be used.}

\item{plot.ncol}{Numeric. A numeric value indicating the number of columns in the plot grid. This setting applies when
more than one data set is provided. The default is set to 1.}

\item{plot.nrow}{Numeric. A numeric value indicating the number of rows in the plot grid. This setting applies when
more than one data set is provided. The default is set to 1.}

\item{plot.output}{Logical. Indicates whether to plot the p-values histograms while the function is running. The default
is set to \code{FALSE}.}

\item{save.se.obj}{Logical. Indicates whether to save the result in the metadata of the SummarizedExperiment object
(\code{se.obj}) or to output the result. By default, it is set to \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE}, messages for different steps of the function will be displayed.}
}
\value{
A SummarizedExperiment object or a list containing the computed AR
}
\description{
This function plots the p-values histograms of the DGE analysis.
}
\details{
Differential expression (DE) analysis is performed using the Wilcoxon signed-rank test with log-transformed data
(e.g., raw counts, normalized data, etc.). To evaluate the effects of different sources of unwanted variation,
DE analysis is performed across batches. In the absence of batch effects, the histogram of the resulting unadjusted
p-values should be uniformly distributed.
}
\author{
Ramyar Molania
}
