% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneSetScore.R
\name{plotGeneSetScore}
\alias{plotGeneSetScore}
\title{Plots the score of the gene set enrichment analysis.}
\usage{
plotGeneSetScore(
  se.obj,
  assay.names = "all",
  reference.score = NULL,
  gene.set.name,
  plot.output = TRUE,
  assess.se.obj = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Character or character vector. Specifies the name(s) of the assay(s) in the SummarizedExperiment
object to calculate sample-wise scores. The default is set to \code{all}, indicating all assays in the SummarizedExperiment
object will be selected.}

\item{reference.score}{Character. The name of the reference score variable.}

\item{gene.set.name}{Character. A character string indicating the name to be used to save the score in the
SummarizedExperiment object. If \code{NULL}, the function will select a name as follows:
\code{gene.set.name <- paste0("singscore|", length(c(upset.genes, downset.genes)), "genes")}}

\item{plot.output}{Logical. Whether to plot the gene set score or not. The default is set to \code{TRUE}.}

\item{assess.se.obj}{Logical. Indicates whether to assess the SummarizedExperiment object or not. The default is set
to \code{TRUE}. Refer to the \code{checkSeObj()} function for more details.}

\item{save.se.obj}{Logical. Indicates whether to save the score results in the metadata of the SummarizedExperiment
object or to output the result as a list. The default is set to \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE}, shows the messages of different steps of the function.}
}
\description{
This function plots the score of the gene set enrichment analysis performed by the \code{computeGeneSetScore()} function.
}
\author{
Ramyar Molania
}
