% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSilhouette.R
\name{plotSilhouette}
\alias{plotSilhouette}
\title{Computes the average Silhouette coefficients.}
\usage{
plotSilhouette(
  se.obj,
  assay.names = "all",
  variables,
  silhouette.method = "sil.euclidian",
  plot.type = "single.plot",
  plot.output = TRUE,
  save.se.obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{assay.names}{Character. A character string or vector of character strings specifying the name(s) of the dataset(s)
in the SummarizedExperiment object to generate barplots or scatter plots of the computed Silhouette coefficients.
By default, all assays in the SummarizedExperiment object will be selected.}

\item{variables}{Character. A character or character vectors of the column names in the SummarizedExperiment object
that the silhouette coefficient has been calculated by the 'computeSilhouette' function.}

\item{silhouette.method}{Character. Indicates which computed Silhouette method should be used for plotting. The method
must match one of those defined in the "computeSilhouette()" function. The default is "sil.euclidian", which is also
the default of the 'computeSilhouette()' function. Refer to the 'computeSilhouette()' function for more details.}

\item{plot.type}{Character. Indicates how to plot the adjusted Silhouette coefficients.
Options are "single.plot" and "combined.plot". If a single variable is provided, "plot.type" must be set to "single.plot",
which generates a barplot of the Silhouette coefficients. If two variables are provided and "plot.type" is set to
"combined.plot", the function generates a scatter plot of the Silhouette coefficients for each variable against each
other. The default is set to 'single.plot'.}

\item{plot.output}{Logical. If TRUE, the individual barplots or scatter plots will be printed during function execution.
The default is set to 'TRUE'.}

\item{save.se.obj}{Logical. Indicates whether to save the plots in the metadata of the SummarizedExperiment object
or to return the results as a list. The default is set to 'TRUE'.}

\item{verbose}{Logical. If TRUE, process messages will be displayed.}
}
\value{
A SummarizedExperiment object or a list containing all the plots of the computed average Silhouette coefficients
for the categorical variables.
}
\description{
This functions generates barplots of average Silhouette coefficients for individual assays. If two variables are
provided, the function creates combined scatter plots of the average Silhouette coefficients of each variables for the
individual assays.
}
\references{
*Molania R., ..., Speed, T. P., Removing unwanted variation from large-scale RNA sequencing data with PRPS,
Nature Biotechnology, 2023
}
\author{
Ramyar Molania
}
