% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVariables.R
\name{plotVariables}
\alias{plotVariables}
\title{Plot variables in a SummarizedExperiment object.}
\usage{
plotVariables(
  se.obj,
  x.variables,
  y.variables,
  cat.test = "anova",
  cont.test = "spearman",
  generate.heatmap = FALSE,
  legend.title = NULL,
  legend.labels = NULL,
  nb.ncol = 4,
  x.lab = NULL,
  y.lab = NULL,
  plot.out.put = TRUE,
  color.palette = "nrc",
  check.se.object = FALSE,
  remove.na = "sample.annotation",
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{TTT}

\item{x.variables}{TTT}

\item{y.variables}{TTTT}

\item{cat.test}{TTTT}

\item{cont.test}{TTTT}

\item{generate.heatmap}{TTTT}

\item{legend.title}{TTTT}

\item{legend.labels}{TTTT}

\item{nb.ncol}{TTTT}

\item{x.lab}{TTTT}

\item{y.lab}{TTTTT}

\item{plot.out.put}{TTTTT}

\item{color.palette}{TTTTT}

\item{check.se.object}{TTTTT}

\item{remove.na}{TTTTT}

\item{verbose}{TTTTT}
}
\description{
This function plots any variables in a SummarizedExperiment object
}
\author{
Ramyar Molania
}
