% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVariables.R
\name{plotVariables}
\alias{plotVariables}
\title{Plots variables in a SummarizedExperiment object.}
\usage{
plotVariables(
  se.obj,
  x.variables,
  y.variables,
  cat.test = "anova",
  cont.test = "spearman",
  generate.heatmap = FALSE,
  check.se.object = TRUE,
  remove.na = "sample.annotation",
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{TTT}

\item{x.variables}{TTT}

\item{y.variables}{TTTT}

\item{cat.test}{TTTT}

\item{cont.test}{TTTT}

\item{generate.heatmap}{TTTT}

\item{check.se.object}{TTTTT}

\item{remove.na}{TTTTT}

\item{verbose}{TTTTT}
}
\description{
This function plots any variables in a SummarizedExperiment object
}
\author{
Ramyar Molania
}
