% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectGenesForPPcorr.R
\name{selectGenesForPPcorr}
\alias{selectGenesForPPcorr}
\title{Selects genes for partial correlation and gene scoring analysis.}
\usage{
selectGenesForPPcorr(
  se.obj,
  assay.names,
  variables,
  groups = NULL,
  cor.cutoff = 0.5,
  abs.cor = TRUE,
  top.genes = NULL,
  method = "spearman",
  a = 0.05,
  rho = 0,
  apply.log = TRUE,
  pseudo.count = 1,
  check.se.object = TRUE,
  remove.na = "both",
  verbose = TRUE
)
}
\arguments{
\item{se.obj}{TTT}

\item{assay.names}{TTT}

\item{variables}{TTTT}

\item{groups}{TTTTsss}

\item{cor.cutoff}{TTTT}

\item{abs.cor}{TTTT}

\item{top.genes}{TTTTT}

\item{method}{TTTTT}

\item{a}{TTTTT}

\item{rho}{TTTTT}

\item{apply.log}{TTTTT}

\item{pseudo.count}{TTTTT}

\item{check.se.object}{TTTTT}

\item{remove.na}{TTTTT}

\item{verbose}{TTTTT}
}
\description{
This function adds the TCGA RNA-seq batch information to the a SummarizedExperiment object.
}
\author{
Ramyar Molania
}
