% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyGenes.R
\name{tidyGenes}
\alias{tidyGenes}
\title{Remove assays from a SummarizedExperiment object}
\usage{
tidyGenes(
  se.obj,
  keep.gene.type,
  gene.type.col.name,
  remove.duplicates.ids = TRUE,
  ids.col.name,
  change.row.names = TRUE,
  new.row.names = NULL
)
}
\arguments{
\item{se.obj}{A SummarizedExperiment object.}

\item{keep.gene.type}{TTTT}

\item{gene.type.col.name}{TTTT}

\item{remove.duplicates.ids}{TTTT}

\item{ids.col.name}{TTTT}

\item{change.row.names}{TTTT}

\item{new.row.names}{TTTT}
}
\description{
Remove assays from a SummarizedExperiment object
}
